% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggDataframe.R
\name{aggDataframe}
\alias{aggDataframe}
\title{Aggregates a data frame to a shorter time period}
\usage{
aggDataframe(CRHMdataframe, columns = 1, period = "annual",
  funs = c("mean"), AggFilename = "", startMonth = 10,
  useSecondYear = TRUE, logfile = "")
}
\arguments{
\item{CRHMdataframe}{Required. A valid \pkg{CRHMr} data frame.}

\item{columns}{The columns to be aggregated, not including the datetime. The default is the first column. This can be a vector, i.e. c(1,2,3).}

\item{period}{The period of aggregation. Must be one of \option{hourly}, \option{daily}, \option{monthly}, \option{yearly} (or \option{annual}) or \option{hydro}. Default is \option{yearly}. Multiple-hour aggregation is not yet supported.}

\item{funs}{A character vector containing the function(s) for aggregation. The default is \option{mean}, but can also include \option{min}, \option{max} and \option{sum}. The function(s) will be applied to all of the specified columns}

\item{AggFilename}{Optional. File name for the aggregated data.}

\item{startMonth}{Optional. Starting month, to be used when aggregating by hydrological year.}

\item{useSecondYear}{Optional. Logical. Should the hydrological year be based on the first or second calendar year. In other words would January 1, 2015 be the hydrological year 2014 or 2015? The default is \code{TRUE} (i.e., the hydrological year would be 2015). Note that the Campbell Scientific program SPLIT uses the first calendar year (i.e., the hydrological year would be 2014). To emulate this program, set \code{useSecondYear} to be \code{FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
Returns a data frame with the aggregated values.
}
\description{
CRHM data (observations and outputs) are generally produced at hourly time steps. This function aggregates CRHM data to daily, monthly or yearly values. The data can be aggregated by their maxima, minima, means, sums and any combination of these statistics.
}
\note{
The period of aggregation must be smaller than the time step of the CRHM data. This function does NOT remove \code{NA} values before aggregation. Use \command{na.omit} or one of the infilling functions (\code{\link{interpolate}} or \code{\link{impute}}) if you want to remove missing values.
}
\examples{
Badlake.t.monthly <- aggDataframe(BadLake7376, period='monthly',
columns=1, funs=c('mean'))
Badlake.radiation.daily <- aggDataframe(BadLake7376, period='daily',
columns=c(6,7,8), funs=c('mean'))
}
\author{
Kevin Shook
}
\seealso{
\code{\link{interpolate}} \code{\link{impute}} \code{\link{yearlyPeaks}} \code{\link{hydroYear}}
}

