% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleObs.R
\name{assembleObs}
\alias{assembleObs}
\title{Assembles two CRHM obs data frames}
\usage{
assembleObs(obs1, obs2, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs1}{Required. The first \pkg{CRHMr} data frame of obs values.}

\item{obs2}{Required. The second obs data frame.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the merges values of both data frames. Note that where the datetimes in the data frames are not the same, the merged values will be set to \code{NA_real_}. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function joins two data frames of CRHM obs. The data frames can have different columns (it is assumed that they will), but should have some dates in common. Both data frames must have the same time step.
}
\examples{
\dontrun{
MSC.trh <- assembleObs(MSC.t, MSC.rh)}
}
\author{
Kevin Shook
}
