% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatePrj.R
\name{automatePrj}
\alias{automatePrj}
\title{Automates a CRHM .prj file}
\usage{
automatePrj(inputPrjFile = "", outputPrjFile = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the the CRHM .prj to be converted.}

\item{outputPrjFile}{Optional. If omitted, the input .prj file will be overwritten.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Converts a CRHM .prj file to be able to be run automatically. It adds the settings \code{Auto_Run}, \code{Auto_Exit} and \code{Log_All} to the end of the .prj file, if they are required.
}
\examples{
\dontrun{
result <- automatePrj('Bad Lake 1974-1975.prj','Bad Lake 1974-1975 auto.prj')}
}
\seealso{
\code{\link{runCRHM}}
}
\author{
Kevin Shook
}
