% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeEatoRH.R
\name{changeEatoRH}
\alias{changeEatoRH}
\title{Changes Ea values to RH in a CRHM data frame}
\usage{
changeEatoRH(obs, t.cols = 1, ea.cols = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame of observations.}

\item{t.cols}{Optional. A vector containing the column numbers (not including the datetime) holding the air temperatures. If no columns are specified then the locations of the temperatures are guessed from the column names. The air temperatures must be in \eqn{^\circ}{ }C.}

\item{ea.cols}{Optional. A vector containing the column numbers (not including the datetime) holding the vapour pressures. If no columns are specified then the locations of the vapour pressures are guessed from the column name. The ea values must be in kPa.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, the function returns the original data frame with the ea columns converted to RH values in perceent. If unsuccessful, it returns the value \code{FALSE}.
}
\description{
This function converts values of Ea to RH. Note that the specified obs data frame must contain both Ea and air temperatures. For safety, \pkg{CRHMr} does not permit values of both RH and Ea in a data frame. The names of the variables containing air temperature and ea values must be of the form \code{t.x} and \code{ea.x}, respectively,  where \code{x} is an number, even if the column numbers are specified.
}
\examples{
# First, convert RH values to Ea
BadLake7376.ea <- changeRHtoEa(BadLake7376)
# now convert Ea values back to RH
BadLake7376.rh <- changeEatoRH(BadLake7376.ea)
}
\seealso{
\code{\link{changeRHtoEa}}
}
\author{
Kevin Shook
}
