% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeRHtoEa.R
\name{changeRHtoEa}
\alias{changeRHtoEa}
\title{Changes RH values to Ea in a CRHM dataframe}
\usage{
changeRHtoEa(obs, t.cols = 1, rh.cols = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame of observations.}

\item{t.cols}{Optional. A vector containing the column numbers (not including the datetime) holding the air temperatures. If no columns are specified then the locations of the temperatures are guessed from the column names. The air temperatures must be in \eqn{^\circ}{ }C.}

\item{rh.cols}{Optional. A vector containing the column numbers (not including the datetime) holding the relative humidities. If no columns are specified then the locations of the RH values are guessed from the column names.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, the function returns the original data frame with the RH columns converted to ea values in kPa. If unsuccessful, it returns the value \code{FALSE}.
}
\description{
This function converts values of RH to Ea. \pkg{CRHMr} does not allow interpolation or imputation of RH values, so you must convert RH values to Ea before infilling. Note that the specified obs dataframe must contain both RH and air temperatures. For safety, \pkg{CRHMr} does not permit values of both RH and Ea in a data frame. The names of the variables containing air temperature and RH values must be of the form \code{t.x} and \code{rh.x}, respectively, where \code{x} is an number, even if the column numbers are specified.
}
\examples{
BadLake7376.ea <- changeRHtoEa(BadLake7376)
}
\author{
Kevin Shook
}
\seealso{
\code{\link{changeEatoRH}}
}

