% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDPtoRH.R
\name{convertDPtoRH}
\alias{convertDPtoRH}
\title{Converts dew point temperature and ambient temperature to RH}
\usage{
convertDPtoRH(t, td)
}
\arguments{
\item{t}{Required. Air temperature in \eqn{^\circ}{ }C.}

\item{td}{Required. Dew point temperature in \eqn{^\circ}{ }C.}
}
\value{
Returns the RH as a percentage.
}
\description{
This is an internal \pkg{CRHMr} function and should normally not need to be called directly. Note that it does \emph{NO} data checking.
}
\examples{
convertDPtoRH(10, 9)
}
\author{
Kevin Shook
}
\references{
Lawrence, M. 2005. The relationship between relative humidity and the dewpoint temperature in moist air. American Meteorological Society: 225-233. \url{http://andrew.rsmas.miami.edu/bmcnoldy/Humidity.htm}
}

