% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRHM_summary.R
\name{CRHM_summary}
\alias{CRHM_summary}
\title{Summarizes a \pkg{CRHMr} data frame}
\usage{
CRHM_summary(CRHMdataframe)
}
\arguments{
\item{CRHMdataframe}{Required. Name of the \pkg{CRHMr} data frame to be summarized.}
}
\value{
If successful, returns a data frame containing the number of rows,
complete rows, starting and ending datetimes, and the variables in the specified
data frame. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Summarizes the values in a \pkg{CRHMr} data frame. This is an internal
\pkg{CRHMr} function and should \emph{never} need to be called directly.
}
\examples{
\dontrun{
summary <- CRHMsummary(BadLake7376)}

}
\author{
Kevin Shook
}
\keyword{internal}
