% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRHM_vars.R
\docType{data}
\name{CRHM_vars}
\alias{CRHM_vars}
\title{CRHM variables}
\format{A data frame with 674 rows and 5 variables:
\describe{
  \item{name}{name of the variable}
  \item{type}{type of variable (its dimensions)}
  \item{description}{description of the variable}
  \item{units}{variable units}
  \item{end_of_day}{\code{TRUE} if the value is only output at the end of the day (00:00), otherwise \code{FALSE}}
}}
\source{
The variable information was abstacted from the \code{CRHM} c++ source code.
}
\usage{
CRHM_vars
}
\description{
A data frame containing information about all of the variables in CRHM that are likely to be output from a model run.
}
\keyword{datasets}

