% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsumDataframe.R
\name{cumsumDataframe}
\alias{cumsumDataframe}
\title{Cumulative sums for a data frame}
\usage{
cumsumDataframe(df)
}
\arguments{
\item{df}{Required. Data frame to be summed.}
}
\value{
Returns a dataframe containing the cumulative sums of all columns in the original data frame.
}
\description{
Finds the cumulative sum of all columns in a data frame. Note that all columns must be numeric - date and datetime variables cannot be summed.
}
\note{
This function is used by other \pkg{CRHMr} functions, so it does \emph{NO} parameter testing.
}
\examples{
\dontrun{
cumul <- cumsumDataframe(modelOutput)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{cumulDailyWater}}
}

