% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeDischargePlot.R
\name{cumulativeDischargePlot}
\alias{cumulativeDischargePlot}
\title{Plots cumulative CRHM and/or HYDAT flows}
\usage{
cumulativeDischargePlot(CRHMflows = NULL, CRHMflowsLabel = "",
  CRHMflowCol = 1, HYDATflows = NULL, HYDATflowsLabel = "",
  facetCols = 3, quiet = TRUE)
}
\arguments{
\item{CRHMflows}{Optional. Optional. A data frame of CRHM modelled flows. The flows must be in m\eqn{^3}{^3}/s.}

\item{CRHMflowsLabel}{Optional. Labels for the CRHM data. If not specified, and CRHM data are plotted, then the name(s) of the CRHM variable(s) will be used.}

\item{CRHMflowCol}{Optional. Column containing the flowrates, not including the datetime. Default is 1.}

\item{HYDATflows}{Optional. Data frame containing WSC daily flows. The data frame is the same as is returned by the function \code{DailyHydrometric} in the package \pkg{HYDAT} developed by David Hutchinson. The data frame has the columns \code{STATION_NUMBER}, \code{DATE}, \code{VALUE} and \code{FLAG}. The \code{DATE} must be an \R date.}

\item{HYDATflowsLabel}{Optional. Labels for the daily flows. If not specified, then the value in \code{STATION_NUMBER} will be used, followed by 'daily'.}

\item{facetCols}{Optional. Number of columns to wrap the facets (if they are used). Setting a value less than 1 will cause an error.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}
}
\value{
If successful, returns a \pkg{ggplot2} object containing the cumulative discharge plot. If unsuccessful, the value \code{FALSE} will be returned.
}
\description{
Creates a \pkg{ggplot} object of annual cumulative flows from data frames of CRHM and/or HYDAT data. The HYDAT flows can be obtained using the \pkg{HYDAT} package created by David Hutchinson. The HYDAT data is truncated so that it only includes the range of the CRHM data. If more than a single year of data is specified, then the plot will be faceted by year.
}
\examples{
\dontrun{
p <- cumulativeDischargePlot(crhm, 'CRHM', 1, dailyflows, 'HYDAT', facetCols=4, quiet=FALSE)}
}
\seealso{
\code{\link{hydrograph}}
}
\author{
Kevin Shook
}
