% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyByHourlyP.R
\name{dailyByHourlyP}
\alias{dailyByHourlyP}
\title{Distributes daily precipitation according to hourly values}
\usage{
dailyByHourlyP(dailyObs, dailyPcol = 0, hourlyObs, hourlyPcol = 0,
  missingRatio = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{dailyObs}{obs Required. A \pkg{CRHMr} data frame containing daily precipitation values.}

\item{dailyPcol}{Optional. The number of the column containing the daily precipitation. If not
specified, the column will be guessed from the variable name. Note that the variable name MUST
include the letter 'p', even if the number is specified.}

\item{hourlyObs}{obs Required. A \pkg{CRHMr} data frame containing hourly precipitation values.}

\item{hourlyPcol}{Optional. The number of the column containing the hourly precipitation. If not
specified, the column will be guessed from the variable name. Note that the variable name MUST
include the letter 'p', even if the number is specified.}

\item{missingRatio}{Optional. The value of the ratio used to multiply the hourly preciptiation when
daily precipitation data are not available. The default is 1. If you want to omit these values,
use \code{NA_real_}.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this
function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are
working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If sucessful, returns a modified version of the obs data frame, consiting of the original
hourly precipitation data, and the \code{adjusted_precip}.
}
\description{
Used to distrbute daily percipitation totals, according to hourly values. Typically,
this is used when the hourly values are of lower quality than the daily values, such as when
they are imputed from another location.
}
\examples{
\dontrun{ hourly <- dailyByHourlyP(dailyObs = daily, hourlyObs = hourly)
}
}
\author{
Kevin Shook
}
