% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFlatlines.R
\name{deleteFlatLines}
\alias{deleteFlatLines}
\title{Deletes values in a obs dataframe column that have a flatline}
\usage{
deleteFlatLines(obs, colnum = 1, window_size = 5, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs data frame.}

\item{colnum}{Optional. The number of the column to test for spikes, not
including the datetime.}

\item{window_size}{Optional. The min number of values within the vector
required to be the same before is flagged as a flat line.}

\item{logfile}{Optional. Name of the file to be used for logging the action.
Normally not used.}
}
\value{
obs dataframe with col flat lines replaced with NANs
}
\description{
Deletes values in a obs dataframe column that have a flatline
}
\examples{
deleteFlatLines(BadLake7376, colnum = 3)

}
\author{
Alex Cebulski
}
