% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePrjHRUs.R
\name{deletePrjHRUs}
\alias{deletePrjHRUs}
\title{Removes specified HRUs from .prj file}
\usage{
deletePrjHRUs(
  inputPrjFile = NULL,
  deletedHRUs = NULL,
  outputPrjFile = NULL,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file to read.}

\item{deletedHRUs}{Required. Vector of HRU numbers to delete.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling
this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default).
If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Deletes names and parameter values for the specified set of HRUs. Note
that this function is very destructive, so it's a good idea to have a backup copy
of your \code{.prj} file \emph{before} running it!
}
\note{
Currently, this function has only been tested on simple \code{.prj} files, i.e.
without REW Grp parameters.
}
\examples{
\dontrun{
result <- deletePrjParameters("CRHM/Bad74_Frozen.prj", c(3, 5, 7)}

}
\seealso{
\code{\link{replacePrjParameters}}
}
\author{
Kevin Shook
}
