% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaStorage.R
\name{deltaStorage}
\alias{deltaStorage}
\title{Calculates changes in storage over specified intervals}
\usage{
deltaStorage(CRHMdataframe = NULL, columns = 1, period = "yearly",
  startMonth = 10, useSecondYear = TRUE, logfile = "")
}
\arguments{
\item{CRHMdataframe}{Required. A standard CRHM data frame.}

\item{columns}{Required. The columns to be differenced, not including the
datetime. The default is the first column. The columns can be a single
number or a vector, i.e. c(1,2,3).}

\item{period}{Required. The period of aggregation. Must be one of \option{hourly},
\option{daily}, \option{monthly}, \option{yearly} (or \option{annual}) or
\option{hydro}. Default is \option{yearly}.}

\item{startMonth}{Optional. Starting month, to be used when differencing by
hydrological year. Default is 10 (October).}

\item{useSecondYear}{Optional. Logical. Should the hydrological year be based on
the first or second calendar year. In other words would January 1, 2015 be the
 hydrological year 2014 or 2015? The default is \code{TRUE} (i.e., the
 hydrological year would be 2015).}

\item{logfile}{Optional. Name of the file to be used for logging the action.
Normally not used.}
}
\description{
Calculates changes in storage over specified intervals
}
\note{
The period of diferencing must be greater than the time step of the
CRHM data. This function does NOT remove \code{NA} values before aggregation.
Use \command{na.omit} or one of the infilling functions (\code{\link{interpolate}} or
\code{\link{impute}}) if you want to remove missing values.
}
\examples{
\dontrun{delta <- deltaStorage(crhm, c(23,24,25))}
}
\seealso{
\code{\link{simpleDailyWater}}
}
\author{
Kevin Shook
}
