% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emissivity}
\alias{emissivity}
\title{Calculates atmospheric emissivity}
\usage{
emissivity(longwave, tK)
}
\arguments{
\item{longwave}{Required. Longwave radiation in W/m\eqn{^2}{^2}.}

\item{tK}{Required. Air temperatures in K.}
}
\value{
Returns the atmospheric emissivity.
}
\description{
Calculates atmospheric emissivity from specified longwave radiation and air
temperatures. This is an internal \pkg{CRHMr} function and should \emph{never} need
to be called directly, as it does \emph{NO} data checking.
}
\examples{
emissivity(100, 293)
}
\seealso{
\code{\link{distributeQli}} \code{\link{longwave}}
}
\author{
Kevin Shook
}
\keyword{internal}
