% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fakeDateHydroyear}
\alias{fakeDateHydroyear}
\title{Calculates fake date for hydrological year}
\usage{
fakeDateHydroyear(dates, fakeYear = 2000, startMonth = 10)
}
\arguments{
\item{dates}{Required. Dates to be converted/}

\item{fakeYear}{Required. Base year.}

\item{startMonth}{Required. Start month for hydrological year.}
}
\value{
Returns vector of fake dates
}
\description{
Calculates fake date for hydrological year
}
\keyword{internal}
