% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fakeDatetime}
\alias{fakeDatetime}
\title{Calculates fake datetime}
\usage{
fakeDatetime(datetime, fakeYear = 2000)
}
\arguments{
\item{datetime}{Required. Datetime to be converted}

\item{fakeYear}{Required. Base year}
}
\value{
Returns vector of fake datetimes
}
\description{
Calculates fake datetime
}
\keyword{internal}
