% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGaps.R
\name{findGaps}
\alias{findGaps}
\title{Finds gaps in obs variables}
\usage{
findGaps(obs, gapfile = "", minlength = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame containing obs values.}

\item{gapfile}{Optional. The name of the output file. If omitted the output file will be the name of the obs data frame, followed by \option{_gaps.csv}.}

\item{minlength}{The minimum gap length included in the analysis (in time steps). Default is 1 time step.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, and there are gaps, returns \code{TRUE}. If successful, and there are no gaps, returns '\option{No gaps}. If unsuccessful, returns \code{FALSE}.
}
\description{
Finds successive missing values in an obs data frame. This function is useful to show where data are missing, before you use the \code{\link{interpolate}} or \code{\link{impute}} functions. All of the gaps in each variable are written to a .csv file.
}
\note{
If quiet=FALSE, the functions gives a summary of the gaps (number of gaps, their total length) for each year.
}
\examples{
findGaps(BadLake7376, quiet=FALSE)
}
\seealso{
\code{\link{insertMissing}} \code{\link{interpolate}} \code{\link{impute}}
}
\author{
Kevin Shook
}
