% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpikesStdevWindow.R
\name{findSpikesStdevWindow}
\alias{findSpikesStdevWindow}
\title{Finds spikes using rolling window stdev filtering}
\usage{
findSpikesStdevWindow(
  obs,
  colnum = 1,
  min_frac_records = 0.8,
  lead_window = list(1:10),
  lag_window = list(-1:-10),
  number_sd = 10,
  logfile = ""
)
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs data frame.}

\item{colnum}{Optional. The number of the column to test for spikes, not
including the \code{datetime}.}

\item{min_frac_records}{Optional. The fraction of records required in a
window for a sucessful computation, otherwise the current value is flagged.
Default is 0.8.}

\item{lead_window}{Optional. A list of values corresponding to the offset
from the current value. Default is 1:10.}

\item{lag_window}{Optional. A list of values corresponding to the offset from
the current value. Default is 1:10.}

\item{number_sd}{Optional. The number of standard deviations away from the
mean required for the current value to be flagged.}

\item{logfile}{Optional. Name of the file to be used for logging the action.
Normally not used.}
}
\value{
If successful and there are no spikes, returns \code{0}. If there are
  spikes, returns their \code{datetime} values. If unsuccessful returns
  \code{FALSE}.
}
\description{
Finds spikes using a rolling window with a width defined as a
  lead and lag offset from the current value. Currently is set up to define
  spikes within the window given a number of standard deviations away from
  the mean. NOTE: the first few and last few values will automatically be
  removed due to the stdev function requiring a minimum of 3 vals.
}
\examples{
findSpikesStdevWindow(BadLake7376, 1, min_frac_records = 0.5)
}
\seealso{
\code{\link{deleteSpikes}} \code{\link{findGaps}}
  \code{\link{findDupes}}
}
\author{
Alex Cebulski
}
