% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMissing.R
\name{insertMissing}
\alias{insertMissing}
\title{Inserts missing datetimes}
\usage{
insertMissing(obs, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame containing obs values.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave quiet=\code{TRUE} (i.e. the default). If you are working interactively, you will probably want to set quiet=\code{FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful returns a modified version of the obs data frame with missing datetime values  inserted. All of the missing variables in the inserted rows are replaced by \code{NA_real_} values. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Many time series have missing rows. This function finds missing rows, and adds rows with the appropriate datetime values and \code{NA_real_} values for all variables. Once the missing values have been added, you can use \code{\link{interpolate}} or \code{\link{impute}} to infill them.
}
\note{
This function should be used before doing interpolation or imputation of variables. Note that this function is also called by the function \code{\link{findGaps}}.
}
\examples{
BadLake.withmissing <- insertMissing(BadLake7376, quiet=FALSE)
}
\seealso{
\code{\link{findGaps}} \code{\link{interpolate}} \code{\link{impute}}
}
\author{
Kevin Shook
}
