% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepGood.R
\name{keepGood}
\alias{keepGood}
\title{Keeps rows where there are some obs values}
\usage{
keepGood(obs, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs dataframe.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the dataframe containing all rows with any non-missing values. If unsuccessful (including if there are only missing values in the original data), returns \code{FALSE}.
}
\description{
Removes rows where all values are missing, keeps the remainder.
}
\examples{
BadLake.good <- keepGood(BadLake7376)
}
<<<<<<< HEAD
<<<<<<< HEAD
\seealso{
\code{\link{trimObs}}
}
\author{
Kevin Shook
}
=======
\author{
Kevin Shook
}
\seealso{
\code{\link{trimObs}}
}

>>>>>>> 38e687f7be6412517ba30e505d371543fddf3102
=======
\seealso{
\code{\link{trimObs}}
}
\author{
Kevin Shook
}
>>>>>>> 885648775cd4bc033f311b418452a30aac53dc41
