% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAction.R
\name{logAction}
\alias{logAction}
\title{Writes a comment to the \pkg{CRHMr} log file}
\usage{
logAction(comment = "", logfile = "")
}
\arguments{
\item{comment}{Required. Comment string to be written. This normally includes the function name, as well as the obs file being processed.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used. If not specified, then 'CRHMr.log' in the current working directory will be used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
This is an internal \pkg{CRHMr} function and should normally not need to be called directly. This function is called by almost all \pkg{CRHMr} functions, and writes the comment, and the date and time to the logfile.
}
\examples{
\dontrun{
 comment <- paste('tMinMaxToHourly dataframe:', obsName, sep='')  
 result <- logAction(comment)}
}
\author{
Kevin Shook
}

