% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{longwave}
\alias{longwave}
\title{Calculates incoming longwave radiation}
\usage{
longwave(em, tK)
}
\arguments{
\item{em}{Required. Emissivity (0-1).}

\item{tK}{Required. Air temperatures in K.}
}
\value{
Returns the longwave radiation in W/m\eqn{^2}{^2}.
}
\description{
Calculates the incoming longwave radiation from the specified emissivity and air temperatures. This is an internal \pkg{CRHMr} function and should \emph{never} need to be called directly, as it does \emph{NO} data checking.
}
\examples{
longwave(0.2, 293)
}
\seealso{
\code{\link{distributeQli}} \code{\link{emissivity}}
}
\author{
Kevin Shook
}
