% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyPrecipTotals.R
\name{monthlyPrecipTotals}
\alias{monthlyPrecipTotals}
\title{Returns mean monthly precipitation totals}
\usage{
monthlyPrecipTotals(obs, precipCols = 0, plot = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame containing \code{p} and/or \code{ppt} values..}

\item{precipCols}{Optional. A vector containing the columns to be analyzed, not including the \code{datetime} column. If not specified, all precipitation columns are used.}

\item{plot}{Optional. If \code{TRUE}, which is the default, a \pkg{ggplot2} object of the monthly mean precipitations is returned. If \code{FALSE}, the values are returned as a data frame.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns either a data frame or a \pkg{ggplot2} object of the monthly mean precipitations. If unsuccessful, returns \code{FALSE}.
}
\description{
Calculates mean total precipitation for each calendar month in the observation data frame. Optionally plots the values. Can work with \code{p} and/or \code{ppt} values.
}
\examples{
\dontrun{
marmot_monthly_plot <- monthlyPrecipTotals(marmot)
marmot_monthly_means <- monthlyPrecipTotals(marmot, plot=FALSE)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{aggDataframe}} \code{\link{monthlyQQplot}}
}

