% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nancumsum.R
\name{nancumsum}
\alias{nancumsum}
\title{Finds cumulative sum of non-missing values in an array}
\usage{
nancumsum(x)
}
\arguments{
\item{x}{Required. An array or vector of numeric values}
}
\value{
Returns an array or vector with cumulative sums of each column.
}
\description{
Finds cumulative sum of non-missing values in an array
}
\examples{
a <- nancumsum(BadLake7376[,2])
}
\author{
Kevin Shook. The code is copied from a MATLAB program nancumsum.m written by Alan Barr.
}
\seealso{
\code{\link{PcpFiltPosTh}}
}

