% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseText.R
\name{parseText}
\alias{parseText}
\title{Parses a string containing several sub-strings}
\usage{
parseText(textString)
}
\arguments{
\item{textString}{Required. A character string containing strings separated by
any number of spaces and/or tabs.}
}
\value{
Returns a vector of individual strings.
}
\description{
This is an internal \pkg{CRHMr} function and should normally not
need to be called directly. It is used for reading data from messy files such
as CRHM projects.
}
\examples{
vals <- parseText(' red  green    blue      black')

}
\author{
Kevin Shook
}
\keyword{internal}
