% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseCorrect.R
\name{phaseCorrect}
\alias{phaseCorrect}
\title{Corrects precipitation for undercatch and calculates phase}
\usage{
phaseCorrect(obs, Tcol = 1, RHcol = 2, Ucol = 3, Pcol = 4,
  RH_type = 1, shield = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} data frame of observations.}

\item{Tcol}{Required. Column number containing the air temperature (not including the datetime). Default is column 1. The values must be in  \eqn{^\circ}{ }C.}

\item{RHcol}{Required. Column number containing the RH (not including the datetime). Default is column 2. The values must be as percentages.}

\item{Ucol}{Required. Column number containing the wind speeds (not including the datetime). Default is column 3. The values must be in m/s.}

\item{Pcol}{Required. Column number containing the precipitaion (not including the datetime). Default is column 4. The values must be in mm.}

\item{RH_type}{Optional. Set to \code{1} if RH is relative to water, any othe value if relative to ice. Default is \code{1}.}

\item{shield}{Optional. Set to \code{1} to use Alter shield undercatch correction from MacDonald and Pomeroy (2007). Set to \code{2} to use Alter shield undercatch correction from Smith. Set to any other value for unadjusted precipitation. Default is \code{1}.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful returns \code{FALSE}. If successful, returns the original dataframe with additional columns for the rain, snow, total precipitation and rain ratio.
}
\description{
Applies an iterative solution to calculate the psychrometric hydrometeor temperature. The hydrometeor temperature is then used to partition the raw precipitation into rainfall and snowfall. The snowfall is adjusted for gauge undercatch, if selected.
}
\examples{
corrected <- phaseCorrect(BadLake7376, Tcol=1, RHcol=2, Ucol=3, Pcol=5)
}
\author{
Phillip Harder
}
\references{
\cite{Harder, P., and J. Pomeroy (2013), \dQuote{Estimating precipitation phase using a psychrometric energy balance method}, Hydrol. Process., 27(13), 1901-1914, doi:10.1002/hyp.9799.} \cr\cr \cite{Macdonald, J., and J. Pomeroy (2007), \dQuote{Gauge Undercatch of Two Common Snowfall Gauges in a Prairie Environment}, Proc. 64th East. Snow Conf. St. John's, Canada., 119-126.}
}

