% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phiv}
\alias{phiv}
\title{Water vapour density (ASHRAE,1993)[gm-3]}
\usage{
phiv(mw, RHi, Tai, Ru)
}
\arguments{
\item{mw}{Unknown}

\item{RHi}{Required. RH as a percent.}

\item{Tai}{Required. Air temp in C.}

\item{Ru}{Unknown}
}
\value{
Returns water vapour density
}
\description{
Water vapour density (ASHRAE,1993)[gm-3]
}
\keyword{internal}
