% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phivt}
\alias{phivt}
\title{Saturated water vapour density (ASHRAE,1993)[gm-3]}
\usage{
phivt(mw, Tai, Ru)
}
\arguments{
\item{mw}{Unknown}

\item{Tai}{Required. Air temperature}

\item{Ru}{Unknown}
}
\value{
Returns saturated water vapour density
}
\description{
Saturated water vapour density (ASHRAE,1993)[gm-3]
}
\keyword{internal}
