% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2runs.R
\name{plot2runs}
\alias{plot2runs}
\title{Plots the results from two CRHM runs}
\usage{
plot2runs(
  run1,
  run2,
  cols = NULL,
  start_datetime = NULL,
  end_datetime = NULL,
  model1_name = "Model 1",
  model2_name = "Model 2"
)
}
\arguments{
\item{run1}{Required. Standard \pkg{CRHMr} data frame of output from first run.}

\item{run2}{Required. Standard \pkg{CRHMr} data frame of output from second run.}

\item{cols}{Variable columns to plot. If omitted, all variables are used. Otherwise
all columns are specified w.r.t. the \code{datetime}.}

\item{start_datetime}{Optional. A POSIXct datetime. If specified, only values >=
the \code{start_datetime} are plotted.}

\item{end_datetime}{Optional. A POSIXct datetime. If specified, only values <=
the \code{end_datetime} are plotted.}

\item{model1_name}{Optional. Name of the first model. Default is \option{Model 1}.}

\item{model2_name}{Optional. Name of the second model. Default is \option{Model 2}.}
}
\value{
Returns ggplot time series plot of the CRHM output, facetted by variable.
}
\description{
Creates a time series plot facetted by variable of values from
two CRHM runs. The same variables must be present in the came columns in
both runs.
}
\examples{
\dontrun{p <- plot2runs("CRHM_run1", "CRHM_run2", cols = c(1, 2, 3))}

}
