% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlags.R
\name{plotFlags}
\alias{plotFlags}
\title{Helper function to plot identified flags}
\usage{
plotFlags(
  obs,
  datetime_flags,
  colnum = 1,
  gg_flag_shape = 4,
  gg_flag_colour = "red",
  gg_flag_size = 1
)
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs data frame.}

\item{datetime_flags}{Required. A vector of datetimes in POSIXct format, for
example the direct output of the findSpikes() function.}

\item{colnum}{Optional. The number of the column to test for spikes, not
including the \code{datetime}.}

\item{gg_flag_shape}{Optional. An integer representing the ggplot shape type
for the flagged values.}

\item{gg_flag_colour}{Optional. A string representing the ggplot colour
for the flagged values.}

\item{gg_flag_size}{Optional. An integer representing the ggplot shape size
for the flagged values.}
}
\value{
a ggplot object.
}
\description{
This function takes a vector of datetimes identified by other
  CRHMr qaqc functions and illustrates these on top of the original data
  using GGPLOT. It is also easy to pass this ggplot output onto plotly for
  interactive viewing by running plotly::ggplotly() after this function.
}
\examples{
plotFlags(BadLake7376, c(as.POSIXct("1974-09-11 12:00:00")))
}
\seealso{
\code{\link{findSpikes}} \code{\link{findGaps}}
  \code{\link{findFlatlines}} \code{\link{findSpikesStdevWindow}}
}
\author{
Alex Cebulski
}
