% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObs.R
\name{plotObs}
\alias{plotObs}
\title{Creates a \pkg{ggplot} object from an obs data frame}
\usage{
plotObs(obs, varcols = c(0), startDate = "", endDate = "", plotType = "lines")
}
\arguments{
\item{obs}{Required. The \pkg{CRHMr} obs data frame to be plotted.}

\item{varcols}{Optional. A vector containing the numbers of the columns (not including the datatime) to be plotted. If not specified, all variables are plotted.}

\item{startDate}{Optional. The starting date for the plot. Can either be a year as a number (e.g. \code{1995}) or a date in \option{Y-m-d} format, i.e. \option{1995-06-01}. If not specified, the first \code{datetime} value in the obs data is used.}

\item{endDate}{Optional. The ending date for the plot. Can either be a year as a number (e.g. \code{1995}) or a date in \option{Y-m-d} format, i.e. \option{1995-12-31}. If not specified, the first \code{datetime} value in the obs data is used.}

\item{plotType}{Optional. Defines the type of the plot. Either \option{lines} or \option{points}. Default is \option{lines}. If you want to show the extent of data, use \option{points} as lines will NOT be drawn when points are separated by \code{NA_real_} values.}
}
\value{
If successful returns a \pkg{ggplot2} object. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function creates a \pkg{ggplot2} object from an obs data frame. If the specified variables are all of single type (e.g. temperature), it returns a single-pane plot. If the specified variables are of more than one type (e.g. temperature, precipitation), then the plot is faceted. The function returns an object that you can modify, or save with the standard \pkg{ggplot2} commands.
}
\details{
Units are automatically specified for obs variables of the following types: \code{t}, \code{rh}, \code{u}, \code{p}, \code{ea}, and solar radiation.
}
\note{
Note that missing (\code{NA_real_}) values in the obs data frame will result in a warning message.
}
\examples{
p <- plotObs(BadLake7376)
print(p)
}
\seealso{
\code{\link{summariseObsFiles}}
}
\author{
Kevin Shook
}
