% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrecipsByYear.R
\name{plotPrecipsByYear}
\alias{plotPrecipsByYear}
\title{Creates a \pkg{ggplot2} object of cumulative precips by year}
\usage{
plotPrecipsByYear(obs = "", obsNames = "", precipCols = 1,
  commonDates = TRUE)
}
\arguments{
\item{obs}{Required. Either a \pkg{CRHMr} \code{obs} data frame or a list of data frames.}

\item{obsNames}{Optional. A vector containing names for the data frames. If there are more data frames than names, the data will be named \code{obs 1}, \code{obs 2} etc.}

\item{precipCols}{Optional. Column(s) containing precipitation. Can be a single value or a vector. Values will be recycled if necessary. Default is \code{1}.}

\item{commonDates}{Optional. If \code{TRUE} (the default) then the range of dates will be restricted to the common first and last dates of the datasets.}
}
\value{
If successful, returns a faceted \pkg{ggplot2} object of cumulative precipitation vs time. If unsuccessful, returns \code{FALSE}.
}
\description{
Calcualtes cumulative precipitations and creates a \pkg{ggplot2} faceted by year. Note that this function would work for any other cumulative variable, if the Y-axis label is changed.
}
\examples{
\dontrun{ l <- list(meas, WRF)
names <- c("Measured", "WRF")
p <- plotPrecipsByYear(l, names)
library(ggplot2)
p <- p + scale_colour_manual(values=c("red", "blue")}
}
\seealso{
\code{\link{plotTempsByYear}}
}
\author{
Kevin Shook.
}
