% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempsByYear.R
\name{plotTempsByYear}
\alias{plotTempsByYear}
\title{Creates a \pkg{ggplot2} object of daily mean air temps by year}
\usage{
plotTempsByYear(obs = "", obsNames = "", tempCols = 1,
  commonYears = TRUE)
}
\arguments{
\item{obs}{Required. Either a \pkg{CRHMr} \code{obs} data frame or a list of data frames.}

\item{obsNames}{Optional. A vector containing names for the data frames. If there are more data frames than names, the data will be named \code{obs 1}, \code{obs 2} etc.}

\item{tempCols}{Optional. Column(s) containing air temperatures. Can be a single value or a scalar. Values will be recycled if necessary.}

\item{commonYears}{Optional. If \code{TRUE} (the default) then the range of years will be restricted to the common first and last years.}
}
\value{
If successful, returns a faceted \pkg{ggplot2} object of mean daily air temperatures vs time. If unsuccessful, returns \code{FALSE}.
}
\description{
Aggregates air temeratures to mean daily values and creates a \pkg{ggplot2} faceted by year. Note that this function would work for any other non-cumulative variable, if the Y-axis label is changed.
}
\examples{
\dontrun{ l <- list(meas, WRF)
names <- c("Measured", "WRF")
p <- plotTempsByYear(l, names)
library(ggplot2)
p <- p + scale_colour_manual(values=c("red", "blue")}
}
