% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qair2ea.R
\name{qair2ea}
\alias{qair2ea}
\title{Converts specific humidity to vapour pressure.}
\usage{
qair2ea(qair, press = 101325)
}
\arguments{
\item{qair}{Required. Specific humidity (dimensionless).}

\item{press}{Optional. Air pressure in Pa. Default is 101325, the standard atmospheric air pressure.}
}
\value{
If successful, returns the vapour pressure in Pa. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Converts specific humidity to vapour pressure.
}
\examples{
ea <- qair2ea(0.0001, 101325)
}
\references{
R code for conversion of air pressure and absolute humidity was taken
from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}.
The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\seealso{
\code{\link{qair2rh}}
}
\author{
Kevin Shook
}
