% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplotValues.R
\name{qqplotValues}
\alias{qqplotValues}
\title{Calculates quantile values for a Quantile-Quantile plot}
\usage{
qqplotValues(x, y)
}
\arguments{
\item{x}{Required. A numeric vector.}

\item{y}{Required. A numeric vector.}
}
\value{
Returns a dataframe with the quantiles of x and y.
}
\description{
The built-in \code{qqplot} function does not work with \pkg{ggplot2}. This just calls the qqplot function to caclulate the quantiles without plotting them.
}
\examples{
quantiles <- qqplotValues(runif(20), runif(50))
}
\author{
Kevin Shook
}
