% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCLASSfile.R
\name{readCLASSfile}
\alias{readCLASSfile}
\title{Reads forcing data from CLASS formatted file}
\usage{
readCLASSfile(
  CLASSfile = "",
  timezone = "",
  convertNames = TRUE,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{CLASSfile}{Required. File of meteorological data in CLASS format.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{convertNames}{Optional. If \code{TRUE}, then the CLASS variables will be converted to \code{Qsi.1}, \code{Qli.1}, \code{p.1}, \code{t.1}, \code{ea.1}, \code{u.1}, and \code{AirPressure.1}. Otherwise, the CLASS names \code{SWin}, \code{LWin}, \code{Prec}, \code{T}, \code{qa}, \code{U}, and \code{P} will be used.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns a \pkg{CRHMr} data frame. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Reads data from files created for CLASS into a stanard \pkg{CRHMr} obs data frame.
}
\examples{
\dontrun{
class <- readCLASSfile('driv.met', timezone='Etc/GMT+7')}
}
\author{
Kevin Shook, with assistance from Andrew Ireson.
}
