% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDebugFile.R
\name{readDebugFile}
\alias{readDebugFile}
\title{Reads in CRHM debug file}
\usage{
readDebugFile(debugFile = "", returnData = "all")
}
\arguments{
\item{debugFile}{Required. CRHM debug file name.}

\item{returnData}{Optional. The data to be returned. If \option{all}
(the default) then all sections will be returned as a list of data frames.
Other options are \option{header}, \option{initalconditions},
\option{basinsummary}, and \option{groupsummary}. Note that if
\option{groupsummary} is specified, and there are no groups, then an
error will result.}
}
\value{
If successful, returns a list of data frames, or a single specified
data frame. If unsuccessful, returns \code{FALSE}.
}
\description{
Reads in CRHM debug file
}
\examples{
/dontrun{
allDebug <- readDebug("07-08_debug.txt")
# show header
View(allDebug$header)
# show group summary
View(allDebug$groupSummary)}
}
