% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjDimensions.R
\name{readPrjDimensions}
\alias{readPrjDimensions}
\title{Reades the dimensions from a .prj file}
\usage{
readPrjDimensions(prjFile = "", logfile = "")
}
\arguments{
\item{prjFile}{Required. Name of .prj file.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the dimension values as a data frame with the columns \code{dimension} and \code{value}.
The dimensions are \code{nhru} (the number of HRUs), \code{nlay} (the number of layers) and \code{nobs} (the
number of obs files). If unsuccessful, the value \code{FALSE} is returned.
}
\description{
Reades the dimensions from a .prj file
}
\examples{
\dontrun{model_dims <- readPrjDimensions("Bad74_Frozen.prj")}

}
\seealso{
\code{\link{readPrjParameters}}
}
\author{
Kevin Shook
}
