% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjHRUnames.R
\name{readPrjHRUnames}
\alias{readPrjHRUnames}
\title{Reads names of HRUs from .prj file}
\usage{
readPrjHRUnames(prjFile = "", logfile = "")
}
\arguments{
\item{prjFile}{Required. Name of .prj file.}

\item{logfile}{logfile Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the parameter values as a vector. If unsuccessful, returns \code{FALSE}.
}
\description{
This function reads names of HRUs from a .prj file. The function \code{readPrjParameters}
will not work, as it only returns numeric values.
}
\examples{
\dontrun{
# read in as a vector
HRUnames <- readPrjHRUnames("Bad_Lake_1974-1975.prj")}
}
\seealso{
\code{\link{setPrjHRUnames}}
}
\author{
Kevin Shook
}
