% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjMatrix.R
\name{readPrjMatrix}
\alias{readPrjMatrix}
\title{Reads maxtrix of specified parameter from .prj file}
\usage{
readPrjMatrix(prjFile = "", paramName = "", logfile = "")
}
\arguments{
\item{prjFile}{Required. Name of .prj file.}

\item{paramName}{Required. Name of parameter to be read.}

\item{logfile}{logfile Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the parameter values as a matrix, with column and row names of the HRU names. If unsuccessful, returns \code{FALSE}.
}
\description{
This function reads numeric parameter values which are in a square matrix where the number of
rows and columns are equal to the number of HRUs.
}
\examples{
\dontrun{
# read in as a matrix
infile <- "Cluster1_PTH_100km2_WFDEI-GEM_new_watershed10_50\%_IN.prj"
variables <- readPrjMatrix(infile, "distrib_Route") }
}
\seealso{
\code{\link{readPrjParameters}} \code{\link{setPrjMatrix}}
}
\author{
Kevin Shook
}
