% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjModuleNames.R
\name{readPrjModuleNames}
\alias{readPrjModuleNames}
\title{Reads names of modules from .prj files}
\usage{
readPrjModuleNames(prjFile, unique_modules = TRUE)
}
\arguments{
\item{prjFile}{Required. Name of .prj file.}

\item{unique_modules}{If \code{TRUE}, only unique module names will be
returned, when there are groups. If \code{FALSE}, each module will be
retuned for each group.}
}
\value{
If successful, returns a data frame containing the columns
\code{module}, \code{type}, and \code{date} for each module. If \option{unique_modules} =
\code{TRUE}, then the variable \code{group} is also returned for each module.
}
\description{
Reads names of modules from .prj files
}
\examples{
\dontrun{
modules <- readPrjModuleNames("Bad74_Frozen.prj")
}
}
\seealso{
\code{\link{readPrjModuleParams}}
}
\author{
Kevin Shook
}
