% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjModuleParams.R
\name{readPrjModuleParams}
\alias{readPrjModuleParams}
\title{Reads all parameters for all HRUs}
\usage{
readPrjModuleParams(moduleName, prjFile)
}
\arguments{
\item{moduleName}{Required. A text pattern of the module name. For the Albedo
Richards module it would be \option{albedo_Richard}.}

\item{prjFile}{Required. Name of .prj file.}
}
\value{
A data frame with the names of the parameters, and the values
of the parameters for each HRU.
}
\description{
Reads all parameters for all HRUs
}
\examples{
\dontrun{
fileName <- "Vuriloche.prj"
shared_parameters <- readPrjModuleParams(moduleName = 'Shared',
prjFile = fileName)}
}
\seealso{
\code{\link{readPrjModuleNames}}
}
\author{
Ezequiel Toum
}
