% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjNumVals.R
\name{readPrjNumVals}
\alias{readPrjNumVals}
\title{Reads numeric values from a CRHM project text object}
\usage{
readPrjNumVals(prj, searchString, value_count)
}
\arguments{
\item{prj}{Required. A CRHM file read into memory as a sequence of lines, as
returned from \code{readPrj}.}

\item{searchString}{Required. The identifying string.}

\item{value_count}{Required. The number of values to be read in.}
}
\value{
If successful, returns a vector with the specified number of values.
If unsuccessful, returns the value \code{FALSE}.
}
\description{
Reads specified number of numeric values from a project file variable.
This is an internal function and is not able to be called by a function outside of
\pkg{CRHMr}, as it is not exported. This documentation is for maintenance purposes.
}
\examples{
\dontrun{
hru_area <- readPrjNumVals(Bologna, 'hru_area', 19)}

}
\keyword{internal}
