% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjOutputVariables.R
\name{readPrjOutputVariables}
\alias{readPrjOutputVariables}
\title{Reads output variables from a CRHM model \code{.prj} file}
\usage{
readPrjOutputVariables(prjFile, asDataframe = FALSE, logfile = "")
}
\arguments{
\item{prjFile}{Required. Name of CRHM .prj file to read.}

\item{asDataframe}{Optional. If \code{FALSE} (the default), returns the set of output variables as a vector. If \code{TRUE}, returns a data frame containing the \code{module}, \code{variable} and \code{HRUs} as a string. The vector formis useful if you want to delete some variables. The data frame is more useful if you want to change the variable or its HRUs.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the output variables. If unsuccessful, returns \code{FALSE}.
}
\description{
Reads output variables from a CRHM model \code{.prj} file
}
\examples{
\dontrun{
# read in as a vector
variables <- readPrjOutputVariables('Bad_Lake_1974-1975.prj')

# read in as a data frame
variables <- readPrjOutputVariables('Bad_Lake_1974-1975.prj', asDataframe=TRUE)
# change value
variables$HRUs[1] <- '1 2'}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{setPrjOutputVariables}}
}

