% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrjParameters.R
\name{readPrjParameters}
\alias{readPrjParameters}
\title{Reads values of specified parameter from .prj file}
\usage{
readPrjParameters(prjFile = "", paramName = "", logfile = "")
}
\arguments{
\item{prjFile}{Required. Name of .prj file.}

\item{paramName}{Required. Name of parameter to be read.}

\item{logfile}{logfile Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the parameter values as a vector. If unsuccessful, returns \code{FALSE}.
}
\description{
This function reads numeric parameter values only. Attempting to read
in character values, sucha s HRU names, will return \code{NULL}.
}
\examples{
\dontrun{
# read in as a vector
variables <- readPrjParameters("Bad_Lake_1974-1975.prj", "Soil soil_moist_init")}
}
\seealso{
\code{\link{setPrjOutputVariables}}
}
\author{
Kevin Shook
}
