% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Regresses one set of CRHM observations against another}
\usage{
regress(primaryCRHM, primary.columns = 1, secondaryCRHM,
  secondary.columns = 1, forceOrigin = FALSE, plot = FALSE,
  quiet = TRUE, logfile = "")
}
\arguments{
\item{primaryCRHM}{Required. The primary \pkg{CRHMr} data frame.}

\item{primary.columns}{Optional. Optional. A vector containing the columns to be imputed in the primary data frame, not including the datetime. If not specified, defaults to the first column.}

\item{secondaryCRHM}{Required. The secondary \pkg{CRHMr} data frame.}

\item{secondary.columns}{Optional. A vector containing the columns to be imputed in the secondary data frame, not including the datetime. If not specified, defaults to the first column.}

\item{forceOrigin}{Optional. If \code{TRUE} then the regressions will be forced through the origin. Note that in this case the values of r\eqn{^2}{^2} and the intercept will \emph{not} be calculated as they have no meaning.}

\item{plot}{Optional. Set \code{plot=TRUE} if you want a plot for each regression. Default is \code{FALSE}. The function has to guess the axis units from the variable names and it may be mistaken; you can always change the labels.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, and plot=\code{TRUE}, returns a plot of all of the regressions. If successful, and plot=\code{FALSE}, returns a data frame containing the  regression parameters. If unsuccessful, returns \code{FALSE}.
}
\description{
The purpose of this function is to generate the multiplier and an offset values (if required) used by the impute function. Therefore, it performs a linear (least-squares) regression of the primary dataset (the values to be produced by imputation) against the secondary dataset. The function removes missing values and merges the data sets together before doing the regression to ensure that the variables are aligned by datetime.
}
\note{
If successful, exports a data frame with the columns \code{xvals} (secondary data frame), \code{yvals} (primary data frame), \code{first date} (in common data frame), \code{last date} (in common data frame), and \code{slope} for each variable to the regression file. If \code{forceOrigin=FALSE} then the parameters \code{intercept}, and \code{r2} will also be returned. This function will not permit regression on RH values, so convert these to Ea values first.
}
\examples{
# this would be used to get the regression coefficients used to impute data from Saskatoon
# to fill gaps in the Regina data
\dontrun{
reg <- regress(regina, c(1,2,3), saskatoon, c(1,2,3))}
}
\seealso{
\code{\link{impute}}
}
\author{
Kevin Shook
}
