% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePrjParameters.R
\name{replacePrjParameters}
\alias{replacePrjParameters}
\title{Replaces the parameter values in a CRHM model \code{.prj} file}
\usage{
replacePrjParameters(
  inputPrjFile = "",
  paramName = "",
  paramVals = "",
  outputPrjFile = "",
  oldParamCount = 0,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{paramName}{Required. Name of parameter to set.}

\item{paramVals}{Required. Vector containing new parameter values.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{oldParamCount}{Optional. If > 0, then this parameter sets the number of
HRUs in the original \code{.prj} file. Otherwise the number of HRUs is read
from the file. The parameter is needed when the number of HRUs in the \code{.prj}
file \code{Dimensions} section does not match the actual number of HRU values
for the parameter.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. I
f you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action.
Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Replaces the existing values of a single parameter, for all HRUs,
in a \code{.prj} file. This differs from \code{setPrjParameters} in that this function can
change the number of values (i.e. the number of HRUs) for a parameter. It
is intended for use in \code{deletePrjHRUs}. Note
that this function is very destructive, so it's a good idea to have a backup copy
of your \code{.prj} file \emph{before} running it!
}
\note{
Currently, this function has only been tested on simple \code{.prj} files,
i.e. without REW Grp parameters. It \emph{might} work for complex files if
the \option{REW Grp} specification is included in the parameter name.
}
\examples{
\dontrun{
result <- replacePrjParameters("CRHM/Bad74_Frozen.prj",
"fetch", c(1500, 100))}
}
\seealso{
\code{\link{deletePrjHRUs}} \code{\link{setPrjParameters}}
}
\author{
Kevin Shook
}
