% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh2vp.R
\name{rh2vp}
\alias{rh2vp}
\title{Converts air temperature and relative humidity to vapour pressure}
\usage{
rh2vp(airtemp, rh)
}
\arguments{
\item{airtemp}{Required. Air temperature in \eqn{^\circ}{ }C.}

\item{rh}{Required. Relative humidity in percent.}
}
\value{
Returns vapour pressure in kPa. If the air temperatures are scalar, returns a scalar value. If the air temperatures are a vector, returns a vector.
}
\description{
This is an internal \pkg{CRHMr} function and should \emph{never} need to be called directly, as it does \emph{NO} data checking. Use the function \code{changeRHtoEa} instead.
}
\examples{
rh2vp(-5, 50)
}
\author{
Kevin Shook
}
\seealso{
\code{\link{changeRHtoEa}}
}

