% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribbonPlotAirTemps.R
\name{ribbonPlotAirTemps}
\alias{ribbonPlotAirTemps}
\title{Plots min, max and mean air temperatures for each day of the year}
\usage{
ribbonPlotAirTemps(obs = "", obsNames = "", tempCols = 1, facet = TRUE,
  facetColour = "", facetCols = 2)
}
\arguments{
\item{obs}{Required. Either a \pkg{CRHMr} \code{obs} data frame or a list of data frames.}

\item{obsNames}{Optional. A vector containing names for the data frames. If there are more data frames than names, the data will be named \code{obs 1}, \code{obs 2} etc.}

\item{tempCols}{Optional. Column(s) containing temperatures. Can be a single value or a vector. Values will be recycled if necessary. Default is \code{1}.}

\item{facet}{Optional. If \code{FALSE} (the default), then all variables will be on a single plot. If \code{TRUE}, then each variable will be on a separate facet.}

\item{facetColour}{Optional. The colour to be used for shading the ribbon plot in each facet. If not specified (the default), then each plot will be coloured separately.}

\item{facetCols}{Optional. The number of columns to be used for the facets, if selected.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a \pkg{ggplot2} object.
}
\description{
Creates \pkg{ggplot2} ribbon plots of the daily minimum, maximum and mean values of air temperatures over a year. The shaded area indicates the range of values in the data frame for each day.
}
\examples{
p <- ribbonPlotAirTemps(BadLake7376, tempCols = 1, facet=FALSE)
}
\seealso{
\code{\link{plotTempsByYear}}
}
\author{
Kevin Shook
}
