% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturatedVP.R
\name{saturatedVP}
\alias{saturatedVP}
\title{Converts calculates saturated vapour pressure from air temperature}
\usage{
saturatedVP(airTemp)
}
\arguments{
\item{airTemp}{Required. Air temperature in \eqn{^\circ}{ }C.}
}
\value{
If successful, returns the saturated vapour pressure in kPa. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This is an internal \pkg{CRHMr} function and should normally not need to be called directly.
}
\examples{
saturatedVP(-5)
}
\author{
Kevin Shook
}
