% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjBasinName.R
\name{setPrjBasinName}
\alias{setPrjBasinName}
\title{Sets the basin name in a CRHM model \code{.prj} file}
\usage{
setPrjBasinName(inputPrjFile = "", basinName = "", outputPrjFile = "",
  logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{basinName}{Required. New name for basin.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets the basin name in a CRHM model \code{.prj} file
}
\examples{
\dontrun{
result <- setPrjBasinName('c:/CRHM/Bad Lake 1974-1975.prj', 'NewBadLakemodel')}
}
\seealso{
\code{\link{runCRHM}} \code{\link{setPrjHRUnames}} \code{\link{setPrjDates}} \code{\link{setPrjParameters}}
}
\author{
Kevin Shook
}
