% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjDates.R
\name{setPrjDates}
\alias{setPrjDates}
\title{Sets the model run start and end dates in a CRHM model \code{.prj} file}
\usage{
setPrjDates(inputPrjFile = "", startDate = "", endDate = "",
  outputPrjFile = "", initialStateFile = "", finalStateFile = "",
  logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{startDate}{Required. Model run starting date in format \option{YYYY MM DD}.}

\item{endDate}{Required. Model run ending date in format \option{YYYY MM DD}.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{initialStateFile}{Optional. If specified, the specified file will be set as the initial state file, i.e. the state variables to be read in at the beginning of the model run.}

\item{finalStateFile}{Optional. If specified, the specified file will be set as the final state file, i.e. the state variables to be exported at the end of the model run.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets the model run start and end dates in a CRHM model \code{.prj} file
}
\note{
If you are using Windows, the paths for the state files must use \option{\\\\} instead of \option{\\} for the directory delimiter. The reason is that R, like many programming languages, interprets \option{\\} as an escape code. This is different from the file path set in \option{inputPrjFile}, which must be an \R path using the \option{/} symbol. See the example below.
}
\examples{
\dontrun{
result <- setPrjDates('c:/CRHM/Bad Lake 1974-1975.prj', 
'1974 10 1', '1975 10 1', initialStateFile='c:\\\\CRHM\\\\Model_Initial_State.int',
 finalStateFile='c:\\\\CRHM\\\\Model_Final_State.int')}
}
\seealso{
\code{\link{runCRHM}}
}
\author{
Kevin Shook
}
