% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjDimensions.R
\name{setPrjDimensions}
\alias{setPrjDimensions}
\title{Sets model dimensions in .prj file}
\usage{
setPrjDimensions(
  inputPrjFile = NULL,
  dimensionVals = NULL,
  outputPrjFile = NULL,
  logfile = ""
)
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file to read.}

\item{dimensionVals}{Required. Vector containing new dimension values. There must be 3 values in the
correct order: \code{nhru} (the number of HRUs), \code{nlay} (the number of layers) and \code{nobs} (the
number of obs files).}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets model dimensions in .prj file
}
\examples{
\dontrun{
new_dimensions <- c(47, 2, 1)
check <- setPrjDimensions("Bad74_Frozen.prj", new_dimensions)}
}
\seealso{
\code{\link{readPrjDimensions}}
}
\author{
Kevin Shook
}
