% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjHRUnames.R
\name{setPrjHRUnames}
\alias{setPrjHRUnames}
\title{Sets the names of HRUs in a CRHM model \code{.prj} file}
\usage{
setPrjHRUnames(inputPrjFile = "", HRUnames = "", outputPrjFile = "",
  logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{HRUnames}{Required. Vector containing new HRU names. There must be at least as many values in this vector as HRU names in the orginal \code{.prj} file.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Replaces the existing names of all HRUs, in a \code{.prj} file.
}
\note{
Currently, this function has only been tested on simple \code{.prj} files, i.e. without REW Grp parameters.
}
\examples{
\dontrun{
result <- setPrjHRUnames('c:/CRHM/Bad74_Frozen.prj', c('fallow', 'stubble', 'grass'))}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{runCRHM}} \code{\link{setPrjParameters}} \code{\link{setPrjDates}}
}

