% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjMatrix.R
\name{setPrjMatrix}
\alias{setPrjMatrix}
\title{Sets the matrix of parameter values in a CRHM model \code{.prj} file}
\usage{
setPrjMatrix(
  inputPrjFile = "",
  paramName = NULL,
  paramMatrix = NULL,
  outputPrjFile = "",
  decimals = 0,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{paramName}{Required. Name of parameter to set.}

\item{paramMatrix}{Required. Square matrix containing new parameter values. There must be at as many rows and
columns in this vector as HRUs in the orginal \code{.prj} file.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{decimals}{Optional. If 0, number of decimal places in output is not set. If >0, number of decimal places
in output is fixed.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function
in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working
interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Replaces the existing values of a single parameter, for all HRUs, in a \code{.prj} file.
}
\note{
Currently, this function has only been tested on simple \code{.prj}
 files, i.e. without REW Grp parameters. It \emph{might} work for complex files if the \option{REW Grp} specification is included in the parameter name.
}
\examples{
\dontrun{
prj_file <- "Cluster1_PTH_100km2_WFDEI-GEM_new_watershed10_50\%_IN.prj"
distrib_route_vals <- readPrjMatrix(prj_file, "distrib_Route")
distrib_route_vals[, 10] <- 0
result <- setPrjMatrix(prj_file, "distrib_Route",distrib_route_vals))}
}
\seealso{
\code{\link{readPrjMatrix}}
}
\author{
Kevin Shook
}
