% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjObs.R
\name{setPrjObs}
\alias{setPrjObs}
\title{Sets the names of the obs files in a CRHM model \code{.prj} file}
\usage{
setPrjObs(inputPrjFile = "", obsFiles = "", outputPrjFile = "",
  logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{obsFiles}{Optional. A string or vector of strings containing the names of the obs file(s). If none are supplied, then the existing obs files will be deleted.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets the names of the obs files in a CRHM model \code{.prj} file
}
\examples{
\dontrun{
obsFiles <- c("BadLakeHourly.obs", "BadLakeppt.obs")
result <- setPrjObs("c:/CRHM/Bad Lake 1974-1975.prj", obsFiles}
}
\seealso{
\code{\link{runCRHM}} \code{\link{setPrjDates}} \code{\link{setPrjParameters}}
}
\author{
Kevin Shook
}
