% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjOutputVariables.R
\name{setPrjOutputVariables}
\alias{setPrjOutputVariables}
\title{Sets the output variables in a CRHM model \code{.prj} file}
\usage{
setPrjOutputVariables(inputPrjFile = "", variables = "",
  outputPrjFile = "", logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{variables}{Required. Variables to be written to output file. These may be either a character vector combining the variable module, variable name and HRUs or a dataframe with the same information in 3 separate columns. Either format can be returned by \code{setOutputVariables}. If the \option{null} is specified for \code{variables}, then all variables are removed.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets the output variables in a CRHM model \code{.prj} file
}
\examples{
\dontrun{
# read in existing variables as a vector
variables <- readPrjOutputVariables('Bad_Lake_1974-1975.prj', asDataframe=FALSE)

# delete the first set of variables and write to a new file
variables <- variables[-1]
setPrjOutputVariables('Bad_Lake_1974-1975.prj', variables, 'Bad_Lake_1974-1975_revised.prj')

# read in existing variables as a dataframe and write to new file
variables <- readPrjOutputVariables('Bad_Lake_1974-1975.prj', asDataframe=TRUE)
newVariables <- c('evap', 'evapGrangerD', '1 2 3 4 5')
variables <- rbind(variables, newVariables)
setPrjOutputVariables('Bad_Lake_1974-1975.prj', variables, 'Bad_Lake_1974-1975_revised2.prj')

# deletes all output variables and overwrites the file
result <- setPrjOutputVariables('Bad_Lake_1974-1975.prj', 'null')}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{readPrjOutputVariables}}
}

