% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjParameters.R
\name{setPrjParameters}
\alias{setPrjParameters}
\title{Sets the parameter values in a CRHM model \code{.prj} file}
\usage{
setPrjParameters(inputPrjFile = "", paramName = "", paramVals = "",
  outputPrjFile = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{paramName}{Required. Name of parameter to set.}

\item{paramVals}{Required. Vector containing new parameter values. There must be at least as many values in this vector as parameter values in the orginal \code{.prj} file.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Replaces the existing values of a single parameter, for all HRUs, in a \code{.prj} file.
}
\note{
Currently, this function has only been tested on simple \code{.prj} files, i.e. without REW Grp parameters. It \emph{might} work for complex files if the \option{REW Grp} specification is included in the parameter name.
}
\examples{
\dontrun{
result <- setPrjParameters('c:/CRHM/Bad74_Frozen.prj',
'fetch', c(1500, 1500, 1500))}
}
\seealso{
\code{\link{runCRHM}} \code{\link{setPrjDates}}
}
\author{
Kevin Shook
}
