% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrjRunID.R
\name{setPrjRunID}
\alias{setPrjRunID}
\title{Sets the Run ID in a CRHM model \code{.prj} file}
\usage{
setPrjRunID(inputPrjFile = "", RunID = 0, outputPrjFile = "",
  logfile = "")
}
\arguments{
\item{inputPrjFile}{Required. Name of the \code{.prj} file.}

\item{RunID}{Required. New RunID for basin. Must be an integer.}

\item{outputPrjFile}{Optional. If omitted, the input \code{.prj} file will be overwritten.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Sets the Run ID in a CRHM model \code{.prj} file
}
\examples{
\dontrun{
result <- setPrjRunID('c:/CRHM/Bad Lake 1974-1975.prj', '5')}
}
\seealso{
\code{\link{runCRHM}} \code{\link{setPrjHRUnames}} \code{\link{setPrjDates}} \code{\link{setPrjParameters}}
}
\author{
Kevin Shook
}
