% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDailyWater.R
\name{simpleDailyWater}
\alias{simpleDailyWater}
\title{Aggregates simple CRHM model output flows to daily values}
\usage{
simpleDailyWater(CRHMoutput, vars = "all", prjFile = "", basinMean = TRUE,
  summarize = TRUE, units = "mm", quiet = TRUE, logfile = "")
}
\arguments{
\item{CRHMoutput}{Required. The CRHM model output as a standard \pkg{CRHMr} dataframe (obs, export or output data)}

\item{vars}{Optional. Variable column numbers to be used (not including the \code{datetime}). The default \option{all} selects all columns.}

\item{prjFile}{Required. The CRHM .prj file.}

\item{basinMean}{Optional. If \code{TRUE} (the default) then the fluxes and storages will be averaged over the entire basin. If \code{FALSE}, then fluxes and storages are computed for each HRU's area.}

\item{summarize}{Optional. If \code{TRUE} (the default), then fluxes and storages are aggregated over the basin  for all HRUs. If \code{FALSE} then fluxes and storages are not aggregated and will be returned for all HRUs.}

\item{units}{Optional. The units for output. The default is \option{mm}, but \option{m3/s} can also be specified.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. The default is TRUE. If you are calling this function in an R script, you will usually leave quiet=TRUE (i.e. the default). If you are working interactively, you will probably want to set quiet=FALSE.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful returns a data frame containing the daily values of all of the water fluxes and storages.
}
\description{
Calculates daily values of the CRHM outputs of water storages and fluxes, for models \emph{without} sub-basins, for each HRU by its area, and calculates the net fluxes for the basin and each sub-basin.
}
\examples{
\dontrun{
b <- readOutputFile('Bologna1984_30y_output.txt', timezone='MST')
# get daily values without aggregation; values calculated w.r.t the HRU area 
dailyHRU <- simpleDailyWater(b, prjFile='Bologna1984_02.prj', basinMean=FALSE)

# get daily values without aggregation; values calculated w.r.t the basin area 
dailyHRU_per_basin <- simpleDailyWater(b, prjFile='Bologna1984_02.prj', basinMean=TRUE)

get daily basin values 
daily_basin <- simpleDailyWater(b, prjFile='Bologna1984_02.prj', basinMean=TRUE, summarize=TRUE)

}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{cumulDailyWater}}
}

