% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMaxSolar.R
\name{simpleMaxSolar}
\alias{simpleMaxSolar}
\title{Estimates extra-terrestrial shortwave radiation}
\usage{
simpleMaxSolar(datetime, latitude, hoursOffset = 2, quiet = TRUE,
  logfile = "")
}
\arguments{
\item{datetime}{Required. Can either be a vector of \pkg{CRHMr} datetimes or any dataframe with a \code{datetime} in column 1.}

\item{latitude}{Required. The latitude for which values are to calculated.}

\item{hoursOffset}{Optional. The offset (in hours) is added to the solar time to convert it to local time. The default value \option{2} shifts the daily peak to occur at 2pm.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns a dataframe containing these variables:
\enumerate{
   \item \code{datetime}
   \item \code{So_MJ}, the daily extra-terrestrial radiation in MJ/m\eqn{^2}{^2}
   \item \code{So_W}, the daily extra-terrestrial radiation in W/m\eqn{^2}{^2}
   \item \code{R_h_d}, the conversion from daily to hourly fluxes
   \item \code{So_h_W}, the sub-daily extra-terrestrial radiation in W/m\eqn{^2}{^2}
   \item \code{So_h_MJ}, the sub-daily extra-terrestrial radiation in MJ/m\eqn{^2}{^2}}
If unsuccessful, the value \code{FALSE} will be returned.
}
\description{
Simple estimation of extra-terrestrial radiation on a horizontal plane. Daily and sub-daily values in MJ/m\eqn{^2}{^2} and W/m\eqn{^2}{^2} are provided,
}
\note{
This version is valid for latitudes between 49 \eqn{^\circ}{ }N and 55\eqn{^\circ}{ }N. It is not known how well it will perform outside this range.
}
\examples{
maxSolar <- simpleMaxSolar(BadLake7376, 51.366)
}
\author{
Kevin Shook
}
\references{
This code is based on Shook, K., and J. Pomeroy (2011), \dQuote{Synthesis of incoming shortwave radiation for hydrological simulation}, Hydrol. Res., 42(6), 433, doi:10.2166/nh.2011.074. \cr Please cite this paper if you use this function in a publication.
}

