% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRibbonPlot.R
\name{simpleRibbonPlot}
\alias{simpleRibbonPlot}
\title{Simple plots of CRHM output max, min and mean values}
\usage{
simpleRibbonPlot(CRHMoutput, columns = "all", waterYear = FALSE,
  ylabel = "Basin total (mm)", skipLeapDay = TRUE, facet = FALSE,
  facetColour = "blue", facetCols = 3)
}
\arguments{
\item{CRHMoutput}{Required. A data frame containing either instantaneous or cumulative values. The data may be either daily or hourly.}

\item{columns}{The columns to be aggregated, not including the datetime. The default is the \option{all}, but can be a vector, i.e. c(1,2,3). Note that all variables must have the same units.}

\item{waterYear}{Optional. If \code{True}, the statistics are calculated and plotted by water year, if \code{FALSE} (the default) the statistics are calculated and plotted by calendar year. If you are plotting cumulative values using \code{cumulDailyWater}, this parameter must be set to be the same as that function.}

\item{ylabel}{Optional. Label to be used for the y-axis.}

\item{skipLeapDay}{Optional. If \code{TRUE} (the default) then values for Feb. 29 will be omitted from the analyses and plots. If \code{FALSE}, then these values will be included, which may result in strange-looking plots.}

\item{facet}{Optional. If \code{FALSE} (the default), then all variables will be on a single plot. If \code{TRUE}, then each variable will be on a separate facet.}

\item{facetColour}{Optional. The colour to be used for shading the ribbon plot in each facet, if selected.}

\item{facetCols}{Optional. The number of columns to be used for the facets, if selected.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a \pkg{ggplot2} object.
}
\description{
Creates \pkg{ggplot2} ribbon plots of the daily minimum, maximum and mean values for all selected variables over a year, for models \emph{without} sub-basins. The shaded area indicates the range of values in the data frame for each day.
}
\examples{
\dontrun{
p <- ribbonPlot(waterYearCumul, waterYear=TRUE, units='mm', facet=TRUE, facetCols=2}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{cumulDailyWater}}
}

