% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseObsFiles.R
\name{summariseObsFiles}
\alias{summariseObsFiles}
\title{Summarises all obs files in a directory}
\usage{
summariseObsFiles(
  file.dir = "",
  timezone = "",
  summaryFile = "obsSummary.csv",
  plot = FALSE
)
}
\arguments{
\item{file.dir}{Optional Directory containing .obs files. If not specified, defaults to current directory. Note that this is an \R path, which uses the \code{'/'} symbol on ALL operating systems.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{summaryFile}{Optional. File to contain summary. Defaults to \option{obsSummary.csv} in the same directory as the obs files.}

\item{plot}{Optional. Logical. Do you want to plot graphs of each obs file? Defaults to \code{FALSE}.}
}
\value{
If successful returns no value. If unsuccessful, returns \code{FALSE}.
}
\description{
This function summarises all of the obs files in specified directory, optionally plotting the values in each obs file in a separate graph.
}
\examples{
\dontrun{
summariseObsFiles('c:/data/BadLake/MSC', timezone='etc/GMT+6', plot=TRUE)}
}
\seealso{
\code{\link{plotObs}}
}
\author{
Kevin Shook
}
