% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestep.hours}
\alias{timestep.hours}
\title{Finds time interval between two datetime values in hours}
\usage{
timestep.hours(datetime1, datetime2)
}
\arguments{
\item{datetime1}{Required. First datetime}

\item{datetime2}{Required. Second datetime}
}
\value{
Returns a numeric value for the difference in hours
}
\description{
Finds time interval between two datetime values in hours
}
\examples{
{
dt1 <- as.POSIXct("2000-01-01 12:00", format = "\%Y-\%m-\%d \%H:\%M")
dt2 <- as.POSIXct("2000-01-01 15:00", format = "\%Y-\%m-\%d \%H:\%M")
timestep.hours(dt1, dt2)}


}
\author{
Kevin Shook Alex Cebulski
}
