% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimObs.R
\name{trimObs}
\alias{trimObs}
\title{Trims a dataframe of missing values at the beginning and end}
\usage{
trimObs(obs, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs dataframe.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the trimmed dataframe. If unsuccessful, returns \code{FALSE}.
}
\description{
Removes values before 01:00 on the first complete day of the data, and after 00:00 on the last complete rows. Because it is destructive, this command should \emph{only} be used immediately before exporting observations to a CRHM .obs file. For daily obs, removes values before the first complete row and after the last complete row.
}
\examples{
BadLake.trimmed <- trimObs(BadLake7376)
}
\author{
Kevin Shook
}
\seealso{
\code{\link{writeObsFile}}
}

