% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorsToVelocity.R
\name{vectorsToVelocity}
\alias{vectorsToVelocity}
\title{Convert wind vectors to velocity}
\usage{
vectorsToVelocity(
  uObs,
  uColnum = 1,
  vObs,
  vColnum = 1,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{uObs}{Required. A \pkg{CRHMr} obs data frame containing the wind u vectors.}

\item{uColnum}{Optional.  The column number containing the u values, not including the datetime. Default is column 1.}

\item{vObs}{Required. A \pkg{CRHMr} obs data frame containing the wind v vectors. Note that u and v must have the same units!}

\item{vColnum}{Optional.  The column number containing the u values, not including the datetime. Default is column 1.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns an obs data frame containing the wind speed (in the original vector units) and the wind direction in degrees from North. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Converts wind u and v vectors to a single wind velocity. The vectors may be in any units, i.e. km/h or m/s.
}
\examples{
\dontrun{
windspeed <- vectorsToVelocity(windObs, uColnum=1, uObs, vColnum=2)}
}
\author{
Kevin Shook
}
