% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp2rh.R
\name{vp2rh}
\alias{vp2rh}
\title{Converts air temperature and vapour pressure to relative humidity}
\usage{
vp2rh(airtemp, vapourPressure)
}
\arguments{
\item{airtemp}{Required. Air temperature in \eqn{^\circ}{ }C.}

\item{vapourPressure}{Required. Vapour pressure in kPa.}
}
\value{
Returns RH in percent.
}
\description{
This is an internal \pkg{CRHMr} function and should \emph{never} need to be called directly, as it does \emph{NO} data checking. Use the function \code{changeEatoRH} instead.
}
\examples{
vp2rh(-5, 0.2007355)
}
\author{
Kevin Shook
}
\seealso{
\code{\link{changeEatoRH}}
}

