% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighingGauge2.R
\name{weighingGauge2}
\alias{weighingGauge2}
\title{Removes spikes from weighing gauge cumulative precipitation}
\usage{
weighingGauge2(obs, precipCol = 1, spikeThreshold = 1000,
  maxSpikeGap = 3, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs file.}

\item{precipCol}{Optional. The number of the column containing the weighting gauge cumulative precipitation data, not including the \code{datetime}. Default is column 1.  This function can only work on 1 column of precipitation.}

\item{spikeThreshold}{Optional. Threshold for single-interval precipitation (mm). Any spikes (changes in cumulative precipitation) whose absoloute value is greater than the threshold will be deleted. The default value is 1000 mm.}

\item{maxSpikeGap}{Optional. Maximum length of spikes (in time steps) to be filled. Default is 3.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If sucessful, returns a modified version of the \code{obs} dataframe containing only the datetime and the adjusted precipitation values.
}
\description{
Removes positive and negative spikes from deaccumulated weighing gauge precipitation. The gaps left from spike removal are infilled by linear interpolation. This function is a wrapper for other functions and was suggested by Craig Smith. This function need not be called if your data does not contain any spikes. It is a good idea to use \code{weighingGaugePlot} to look at your data before calling this function. Note that this function will also remove gauge resets, if they are wihin the range of the \code{spikeThreshold}.
}
\note{
If you don't specify a large enough value for \code{maxSpikeGap}, then all of the values after the spike will be set to \code{NA}.
}
\examples{
\dontrun{
test2 <- weightingGauge2(wg, spikeThreshold = 300)}

}
\seealso{
\code{\link{weighingGauge1}} \code{\link{weighingGauge3}} \code{\link{weighingGauge4}} \code{\link{weighingGaugePlot}} \code{\link{weighingGaugeInterval}}
}
\author{
Kevin Shook
}
