% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighingGauge3.R
\name{weighingGauge3}
\alias{weighingGauge3}
\title{Filters weighing gauge cumulative precipitation}
\usage{
weighingGauge3(obs, precipCol = 1, filterLength = 0, quiet = TRUE,
  logfile = "")
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs file.}

\item{precipCol}{Optional. The number of the column containing the weighing guage cumulative precipitation data, not including the \code{datetime}. Default is column 1.  This function can only work on 1 column of precipitation.}

\item{filterLength}{Required. The number of time intervals used by the Savitzky-Golay polynomial filter. Note that this must be an ODD number. The appropriate value will depend on the type of your data, and the noise present.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If sucessful, returns a modified version of the \code{obs} dataframe containing the adjusted precipitation values. Note that the name of the precipitation variable is modified: \code{'_sg_filtered'} is appended.
}
\description{
Thes functions is a wrapper for the Savitzky-Golay polynomial filter function \code{sgolayfilt} from the package \pkg{signal}. It passes the specified cumulative precipitation data, the filter length, the filter order (set to 1) and the derivative order (set to 0) to the filter. It is a good idea to use \code{weighingGaugePlot} to look at your data before and after calling this function. Note that this function is only intended to be used with very high frequency (i.e. 1 minute) gauged data, and should not be used with longer-period accumulations.
}
\note{
This function is potentially destructive, as it smooths the accumulated precipitation. You may need to set the filter length by trial-and error. You can check the effects using the function \code{weighingGaugePlot}.
}
\examples{
\dontrun{
test3 <- weighingGauge3(wg, filterLength=5)}

}
\seealso{
\code{\link{weighingGauge1}} \code{\link{weighingGauge2}} \code{\link{weighingGauge4}} \code{\link{weighingGaugePlot}} \code{\link{weighingGaugeInterval}}
}
\author{
Kevin Shook
}
