% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighingGauge4.R
\name{weighingGauge4}
\alias{weighingGauge4}
\title{Runs Alan Barr's filter on weighing gauge cumulative precipitation}
\usage{
weighingGauge4(
  obs,
  precipCol = 1,
  smallDropThreshold = 0.1,
  serviceThreshold = -100,
  serviceGapLength = 3,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs file.}

\item{precipCol}{Optional. The number of the column containing the weighing gauge cumulative percipitaion data, not including the \code{datetime}. Default is column 1.  This function can only work on 1 column of precipitation.}

\item{smallDropThreshold}{Optional. Minimum interval precipitation (mm). Default is 0.1.}

\item{serviceThreshold}{Optional. Threshold for removing change in storage due to servicing (mm). Note that this must be a NEGATIVE value. Default is -100.}

\item{serviceGapLength}{Optional. Maximum gap length (in time intervals) to be filled by linear interpolation after removing servicing (i.e. gauge reset) data. The default value is 3, but you may have to adjust it for very high-frequency data.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If sucessful, returns a modified version of the \code{obs} dataframe containing the adjusted precipitation values. Note that the name of the precipitation variable is modified: \code{'_PcpFiltPosT'} is appended.
}
\description{
This function is a wrapper for Alan Barr's function \code{PcpFiltPosTh} which removes jitter from precipitation data. The called function is iterative and may be slow to execute.
}
\note{
This function is potentially destructive, as it smooths the accumulated precipitation. You may need to set the filter length by trial-and error. You can check the effects using the function \code{weighingGaugePlot}. This version does not explicitly incorporate the effects of evaporation.
}
\examples{
\dontrun{
test4 <- weighingGauge4(wg, smallDropThreshold = 0.05, serviceThreshold = -50)}

}
\seealso{
\code{\link{PcpFiltPosTh}} \code{\link{weighingGauge1}} \code{\link{weighingGauge2}} \code{\link{weighingGauge3}} \code{\link{weighingGaugePlot}} \code{\link{weighingGaugeInterval}}
}
\author{
Kevin Shook
}
