% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighingGauge5.R
\name{weighingGauge5}
\alias{weighingGauge5}
\title{Simplified weighing gauge processing}
\usage{
weighingGauge5(obs, precipCol = 1, resetThreshold = 50, quiet = TRUE,
  logfile = "")
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs file.}

\item{precipCol}{Optional. The number of the column containing the weighing gauge cumulative precipitation data, not including the \code{datetime}. Default is column 1. These values are cumulative preciptiation. This function can only work on 1 column of precipitation. The precipitaition cannot contain missing values - infill them using \code{weighingGauge1} first.}

\item{resetThreshold}{Optional. The minimum (absolute value) of negative values considered to be gauge resets. Default is 50 mm.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If sucessful, returns a modified version of the \code{obs} dataframe containing only the datetime and the modified cumulative precipitation values.
}
\description{
This function removes gauge resets (large negative changes usually due to servicing) and uses a simple method (cumulative maxima) to remove jitter. It is a good idea to use \code{weighingGauge1} to infill missing values before calling this function and \code{weighingGaugePlot} to look at your data before and after calling this function.
}
\note{
This version does not explicitly incorporate the effects of evaporation.
}
\examples{
\dontrun{
wg1 <- weighingGauge1(wg,  maxGapLength=500)
wg5 <- weighingGauge5(wg1,  resetThreshold=70)}

}
\seealso{
\code{\link{weighingGauge1}} \code{\link{weighingGaugePlot}} \code{\link{weighingGaugeInterval}}
}
\author{
Kevin Shook, based on an algorithm developed by Thai Nguyen of Alberta Environment.
}
