% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighingGaugeInterval.R
\name{weighingGaugeInterval}
\alias{weighingGaugeInterval}
\title{Deaccumulates weighing gauge precipitation}
\usage{
weighingGaugeInterval(obs, precipCol = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs file.}

\item{precipCol}{Optional. The number of the column containing the weighting gauge cumulative precipitation data, not including the \code{datetime}. Default is column 1.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If sucessful, returns a modified version of the \code{obs} dataframe containing the datetime, the cumulative and interval precipitation.
}
\description{
Converts weighing gauged cumulative precipitation to interval values. The interval precipitation is checked to make sure that it does not contain any negative values. Negative precipitation will result in the function displaying an error message and returning the value \code{FALSE}.
}
\examples{
\dontrun{
test5 <- weightingGaugeStep5(wg)}

}
\seealso{
\code{\link{weighingGauge1}} \code{\link{weighingGauge2}} \code{\link{weighingGauge3}}  \code{\link{weighingGauge4}}  \code{\link{weighingGauge5}} \code{\link{weighingGaugePlot}}
}
\author{
Kevin Shook
}
