% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingGaugePlot.R
\name{weighingGaugePlot}
\alias{weighingGaugePlot}
\title{Plots weighing gauge precipitation}
\usage{
weighingGaugePlot(obs, precipCol = 1, startDate = "", endDate = "",
  showMissing = TRUE)
}
\arguments{
\item{obs}{Required. A standard \pkg{CRHMr} obs dataframe.}

\item{precipCol}{Optional. The column containing the cumulative precipitation data. Default is column \code{1}.}

\item{startDate}{Optional. The starting date for the plot. Can either be a year as a number (e.g. \code{1995}) or a date in \option{Y-m-d} format, i.e. \option{1995-06-01}. If not specified, the first \code{datetime} value in the obs data is used.}

\item{endDate}{Optional. The ending date for the plot. Can either be a year as a number (e.g. \code{1995}) or a date in \option{Y-m-d} format, i.e. \option{1995-12-31}. If not specified, the first \code{datetime} value in the obs data is used.}

\item{showMissing}{Optional. If set to \code{TRUE} (the default), then the position of missing values will be indicated by red points on the x-axis.}
}
\value{
If successful returns a \code{ggplot2} object. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Plots culumative precipitation data from a weighting gauge. The plot contains facets with the cumulative and interval precipitations plotted against time.
}
\examples{
\dontrun{
p <- weighingGaugePlot(test1) }
}
\author{
Kevin Shook.
}
\seealso{
\code{\link{weighingGauge1}} \code{\link{weighingGauge2}} \code{\link{weighingGauge3}} \code{\link{weighingGauge4}}  \code{\link{weighingGauge5}}
}

