% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetland_h_s.R
\name{wetland_h_s}
\alias{wetland_h_s}
\title{Calculates Hayashi-van der Kamp parameters for a Prairie depressions}
\usage{
wetland_h_s(
  max_volume,
  max_area,
  p,
  max_iterations = 100,
  max_error = 1e-04,
  return_stats = FALSE
)
}
\arguments{
\item{max_volume}{Required. Maximum volume of a pond in m\eqn{^3}{^3}.}

\item{max_area}{Required. Maximum area of a pond in m\eqn{^2}{^2}.}

\item{p}{Required. Scaling exponent for pond depth-area relationshp. See reference.}

\item{max_iterations}{Optional. Maximum number of iterations for solution. Default is 100.}

\item{max_error}{Optional. Maximum error for estimation of parameters, as a fraction of \code{max_area}.}

\item{return_stats}{If \code{TRUE}, then the error of solution and the number of iterations
will also be returned. Default is \code{FALSE}
Default is 0.001.}
}
\value{
If successful, returns a vector of the parameters \code{h} (the maximum pond depth) in m and
\code{s} (the pond area for a depth of 1 m) in m\eqn{^2}{^2}). Optionally returns the error and number of iterations
required for solution. If unsuccessful, returns \code{FALSE}.
}
\description{
Computing pond area and depth for Prairie depressions from the Hayashi-van der Kamp method
requires values of the parameters h and s. Unfortunately, these values can't be solved explicitly
from the pond max area and volume. This function finds the values of the parameters by iterative solution.
}
\examples{
wetland_h_s(5000, 500, 1.72)

}
\references{
\cite{Hayashi, M., and G. van der Kamp (2000), Simple equations to represent the volume–area–depth relations
of shallow wetlands in small topographic depressions, J. Hydrol., 237(1–2), 74–85, doi:10.1016/S0022-1694(00)00300-0.}
}
\seealso{
Other wetland_functions: 
\code{\link{wetland_pond_area}()},
\code{\link{wetland_pond_depth}()},
\code{\link{wetland_pond_size}()}
}
\author{
Kevin Shook
}
\concept{wetland_functions}
