% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetland_pond_area.R
\name{wetland_pond_area}
\alias{wetland_pond_area}
\title{Calculates Hayashi-van der Kamp area of water in a Prairie pond}
\usage{
wetland_pond_area(pond_depth = NULL, p = NULL, s = NULL)
}
\arguments{
\item{pond_depth}{Required. Depth of water in pond (in m). May be a vector of values. Note that
in this context the depth refers to the deepest point in the pond.}

\item{p}{Required. Value of \code{p} exponent for ponds. May be a vector of values.}

\item{s}{Required. Value of \code{s} parameter (pond area for a depth of 1 m) in m\eqn{^2}{^2}) for ponds.
May be a vector of values.}
}
\value{
If successful, returns a vector of the pond areas, for the given depths.
If unsuccessful, returns \code{FALSE}.
}
\description{
Calculates Hayashi-van der Kamp area of water in a Prairie pond
}
\examples{
 wetland_pond_area(1, 1.72, 25)
}
\references{
\cite{Hayashi, M., and G. van der Kamp (2000), Simple equations to represent the volume–area–depth relations
of shallow wetlands in small topographic depressions, J. Hydrol., 237(1–2), 74–85, doi:10.1016/S0022-1694(00)00300-0.}
}
\seealso{
Other wetland_functions: \code{\link{wetland_h_s}},
  \code{\link{wetland_pond_depth}},
  \code{\link{wetland_pond_size}}
}
\author{
Kevin Shook
}
\concept{wetland_functions}
