% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetland_pond_size.R
\name{wetland_pond_size}
\alias{wetland_pond_size}
\title{Estimates pond area, depth and perimeter (circular)}
\usage{
wetland_pond_size(
  CRHM_output = NULL,
  HRU_name = NULL,
  HRU_SD_max = NULL,
  HRU_area = NULL,
  wetland_HRU = NULL,
  area_p_vals = NULL,
  logfile = ""
)
}
\arguments{
\item{CRHM_output}{Required. Data frame of CRHM outputs including Sd values for all wetland HRUs.}

\item{HRU_name}{Required. Names of all HRUs. Can be obtaibes using \code{readPrjParameters}.}

\item{HRU_SD_max}{Required. SDmax values for all HRUs. Can be obtaibes using \code{readPrjParameters}.}

\item{HRU_area}{Required. Areas of all HRUs. Can be obtaibes using \code{readPrjParameters}.}

\item{wetland_HRU}{Optional. A vector of the number of each wetland HRUs.}

\item{area_p_vals}{Optional. A data frame containing two columns. The first column is the
a wetland area (in m\eqn{^2}{^2}), the second column is the corresponding value of \code{p} to use.
If this parameter is not specified, then the default values will be used. The default is to use
\code{p = 1.72} for areas greater than 0 and smaller than 10,000 m\eqn{^2}{^2}, and \code{p = 3.3}
for areas greater than or equal to 10,000 m\eqn{^2}{^2}. These values were estimated for
Smith Creek basin.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns a data frame consisting of the \code{datetime} (or \code{date}), and
the pond area (m\eqn{^2}{^2}), depth (m), and perimeter (m) (assuming the pond is circular) for
each time step in \code{CRHM_output}.
}
\description{
Estimates pond area, depth and perimeter (circular)
}
\note{
Because this function may be slow and generates a very large quantity of output, it
is a good idea to restrict the range of \code{CRHM_output} \emph{before} calling the function.
}
\examples{
\dontrun{
CRHM_output <- readOutputFile("CRHM_output_1.txt", "Etc/GMT+6")
HRU_name <- readPrjHRUnames("SmithCreek.prj")
HRU_SD_max <- readPrjParameters("SmithCreek.prj", "Sdmax")
HRU_area <- readPrjParameters("SmithCreek.prj", "hru_area")
area <- c(0, 5000, 10000)
p <- c(1.4, 1.7, 3.3)
area_p_vals <- data.frame(area, p)
pond_sizes <- wetland_pond_size(CRHM_output, HRU_name, HRU_SD_max, HRU_area, area_p_vals)}
}
\seealso{
Other wetland_functions: 
\code{\link{wetland_h_s}()},
\code{\link{wetland_pond_area}()},
\code{\link{wetland_pond_depth}()}
}
\author{
Kevin Shook
}
\concept{wetland_functions}
