% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCLASSfile.R
\name{writeCLASSfile}
\alias{writeCLASSfile}
\title{Writes forcing data to a CLASS file}
\usage{
writeCLASSfile(CLASS, CLASSfile = "", logfile = "")
}
\arguments{
\item{CLASS}{Required. A \pkg{CRHMr} data frame containing the fields  \code{SWin}, \code{LWin}, \code{Prec}, \code{T}, \code{qa}, \code{U}, and \code{P}, in addition to the \code{datetime}.}

\item{CLASSfile}{Required. A the file to hold the CLASS data.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE}. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Writes a standard \pkg{CRHMr} data frame to a CLASS forcing data file.
}
\examples{
\dontrun{
result <- writeCLASSfile(class, 'driv.met')}
}
