% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeObsFile.R
\name{writeObsFile}
\alias{writeObsFile}
\title{Writes a dataframe to a .obs file}
\usage{
writeObsFile(obs, obsfile = "", comment = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs data frame.}

\item{obsfile}{Required. Name for file to hold observations.}

\item{comment}{Optional. A comment to be added to the header of the .obs file. If omitted, the header will consist of the name of the obs data frame followed by \option{created by CRHMr function writeObsFile}.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Writes a CRHM dataframe to an obs file. It will not allow ea values - they must be converted to RH values first. The CRHM filter command to convert RH values to ea is added automatically to the header of the .obs file.
}
\note{
If there are missing values in the obs dataframe, the last line of the header will contain the line \option{$$ Missing}.
}
\examples{
# write sample obs dataframe to .obs file
result <- writeObsFile(BadLake7376, 'BadLake7376.obs')
}
\seealso{
\code{\link{trimObs}} \code{\link{changeEatoRH}}
}
\author{
Kevin Shook
}
