% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPeaks.R
\name{yearlyPeaks}
\alias{yearlyPeaks}
\title{Finds datetimes of annual max values}
\usage{
yearlyPeaks(CRHMdata, columns = 0, minvals = FALSE, logfile = "")
}
\arguments{
\item{CRHMdata}{Required. A \pkg{CRHMr} dataframe.}

\item{columns}{Optional. A vector containing the columns to be analyzed, not including the datetime column. If not specified, all data columns are used.}

\item{minvals}{Optional. Logical. Specifies if yearly minimum values should also be found. Default value is \code{FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
Returns a dataframe with maximum (and minimum if selected) values, and their corresponding datetimes, for each year for each selected variable.
}
\description{
Finds the annual maxima (and, optionally, minima) of the specfied columns in a CRHM dataframe and their corresponding datetimes.
}
\examples{
\dontrun{
p <- yearlyPeaks(BrandonObs, columns=c(1,2), minvals=TRUE)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{aggDataframe}}
}

