% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroMissingPrecip.R
\name{zeroMissingPrecip}
\alias{zeroMissingPrecip}
\title{Sets missing precipitation values to zero}
\usage{
zeroMissingPrecip(obs, precipCols = 0, quiet = TRUE, logfile = "")
}
\arguments{
\item{obs}{Required. A \pkg{CRHMr} obs dataframe.}

\item{precipCols}{Optional. A vector containing the numbers of the columns (excluding the datetime) to be processed. If omitted, all of the columns whose names contain \option{p} or \option{ppt} will be processed}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If the action is successful, returns a dataframe with all missing precipitation values set to zero. If the action is unsuccessful, then the value \code{FALSE} is returned. If \code{quiet=FALSE}, then a summary of the returned obs data will be returned.
}
\description{
Removes missing precipitation (\code{NA_real_}values (p and/or ppt) in an obs dataframe, by setting them to zero. This function is potentially dangerous, as setting the missing values to zero prevents any further imputation/interpolation of values.
}
\note{
As described, this function is potentially dangerous, and its action \emph{CANNOT} be undone. It should only be used when there is no further hope for imputing/interpolating precipitation data. Note that its use is really only justified in arid or semi-arid climates, where precipitation is rare. Because it is so destructive, the function should only be used immediately before exporting to an obs file.
}
\examples{
BadLake7376.clean <- zeroMissingPrecip(BadLake7376, quiet=FALSE) 
}
\author{
Kevin Shook
}
\seealso{
\code{\link{writeObsFile}}
}

