% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeDischargePlot.R
\name{cumulativeDischargePlot}
\alias{cumulativeDischargePlot}
\title{Plots cumulative CRHM and/or WSC flows}
\usage{
cumulativeDischargePlot(
  CRHMflows = NULL,
  CRHMflowsLabel = "",
  CRHMflowCol = 1,
  WSCdailyFlowsID = NULL,
  WSCdailyFlowsLabel = "",
  facetCols = 3,
  quiet = TRUE,
  hydat_path = NULL
)
}
\arguments{
\item{CRHMflows}{Optional. Optional. A data frame of CRHM modelled flows. The flows must
be in m\eqn{^3}{^3}/s.}

\item{CRHMflowsLabel}{Optional. Labels for the CRHM data. If not specified, and CRHM data
are plotted, then the name of the CRHM variable will be used.}

\item{CRHMflowCol}{Optional. Column containing the flowrates, not including the datetime.
Default is 1.}

\item{WSCdailyFlowsID}{Optional. If \code{NULL} (the default) the WSC daily flows will not be
plotted. If a WSC station ID is specified , e.g. \code{WSCdailyFlowsID = "05CC001"}, then
the daily flows will be obtained from \pkg{tidyhydat} and plotted.}

\item{WSCdailyFlowsLabel}{Optional. Labels for the daily flows. If not specified, then the
WSC station number will be used, followed by \option{daily}.}

\item{facetCols}{Optional. Number of columns to wrap the facets (if they are used).
Setting a value less than 1 will cause an error.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling
this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default).
If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{hydat_path}{Optional. Path to the HYDAT database. This can usually be left blank.}
}
\value{
If successful, returns a \pkg{ggplot2} object containing the cumulative discharge
plot. If unsuccessful, the value \code{FALSE} will be returned.
}
\description{
Creates a \pkg{ggplot} object of annual cumulative flows from a data frame of CRHM
output and/or WSC daily flows. The WSC flows are obtained from the \pkg{tidyhydat} package.
The WSC data is truncated so that it only includes the range of the CRHM data.
If more than a single year of data is specified, then the plot will be faceted by year.
}
\examples{
\dontrun{
p <- cumulativeDischargePlot(CRHMflows = crhm_output, WSCdailyFlowsID = "05FA025",
 facetCols=2, quiet=FALSE)}
}
\seealso{
\code{\link{hydrograph}}
}
\author{
Kevin Shook
}
