% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{hydrograph}
\alias{hydrograph}
\title{Plots hydrograph of CRHM output and/or WSC daily and/or peak flows}
\usage{
hydrograph(CRHMflows = NULL, CRHMflowsLabels = "", CRHMcols = NULL,
  CRHMdaily = FALSE, WSCdailyFlows = NULL, WSCdailyFlowsLabel = "",
  WSCpeakFlows = NULL, WSCpeakFlowsLabel = "",
  forceMissingPeakTimes = FALSE, commonTime = FALSE, fakeDates = FALSE,
  quiet = TRUE)
}
\arguments{
\item{CRHMflows}{Optional. A data frame of CRHM modelled flows. The flows must be in m\eqn{^3}{^3}/s.}

\item{CRHMflowsLabels}{Optional. Labels for the CRHM data. If not specified, and CRHM data are plotted, then the name(s) of the CRHM variables will be used.}

\item{CRHMcols}{Required. Column(s) containing the flowrates. As always, the numbers do not include the \code{datetime}.}

\item{CRHMdaily}{Optional. Should CRHM flows be plotted as daily values? Default is \code{FALSE}.}

\item{WSCdailyFlows}{Optional. Dataframe containing WSC daily flows. The data frame is the same as is returned by the function \code{DailyHydrometricData} in the \pkg{HYDAT} package developed by David Hutchinson. The datframe has the columns \code{station_number}, \code{date}, \code{value} and \code{flag}. The \code{date} must be an \R date.}

\item{WSCdailyFlowsLabel}{Optional. Labels for the daily flows. If not specified, then the value in \code{station_number} will be used, followed by \option{daily}.}

\item{WSCpeakFlows}{Optional. Dataframe containing WSC annual peak flows for a single station. The data frame is the same as is returned by the function \code{AnnualPeakData} in the \pkg{HYDAT} package developed by David Hutchinson. The data frame has the columns \code{station_number}, \code{data_type}, \code{year}, \code{peak_code}, \code{precision_code}, \code{month}, \code{day}, \code{hour}, \code{minute}, \code{time_zone}, \code{peak}, and \code{symbol}.}

\item{WSCpeakFlowsLabel}{Optional. Labels for the annual peak flows. If not specified, then the value in \code{station_number} will be used, followed by \option{annual peak}.}

\item{forceMissingPeakTimes}{Optional. Some peaks may be missing their time of day and/or time zone. If \code{TRUE}, the missing peak times will be set to noon on the date of record. The missing time zone will be set to that of the other values or the user's time zone if there are none specified. If \code{FALSE} (the default value) peaks missing times and/or timezones will be deleted from the plot.}

\item{commonTime}{Optional. If set to \code{TRUE} then the hydrographs will only plotted over their common time range. Default is \code{FALSE}.}

\item{fakeDates}{Optional. If set to \code{TRUE} then all dates have their year replaced with \code{2000}, and the actual year is added as a variable in the plotted data. This allows the plot to be faceted by year, as shown in the examples.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}
}
\value{
If successful, returns a \pkg{ggplot2} object. If unsuccessful, returns \code{FALSE}.
}
\description{
Creates a \pkg{ggplot} hydrograph from any of CRHM flows, WSC daily flows and/or WSC peak flows.
}
\note{
The CRHM flows are plotted as lines, the daily flows are plotted as steps, and the annual peaks are plotted as points.
}
\examples{
\dontrun{
p <- hydrograph(BadLakeModel,'CRHM Bad Lake model',1, dailyFlows, '', peakFlows, '')
# once the ggplot graph has been returned, it can easily be modified:
mintime <- as.POSIXct(as.Date('1975-04-01', format='\%Y-\%m-\%d'), tz='etc/GMT+6')
maxtime <- as.POSIXct(as.Date('1975-04-30', format='\%Y-\%m-\%d'), tz='etc/GMT+6')
p <- p + xlim(mintime, maxtime) + ylim(0, 4)

# re-plot with fake dates
p2 <- hydrograph(BadLakeModel,'CRHM Bad Lake model',1, dailyFlows, '', peakFlows, '',
commonTime=TRUE, fakeDates=TRUE)
p2 <- p2 + facet_wrap(~year, scales='free_y', ncol=8)
# set axis limits to be the specified date range, and set labels to be the month names
mintime <- as.POSIXct(as.Date('2000-03-01', format='\%Y-\%m-\%d'), tz='etc/GMT+6')
maxtime <- as.POSIXct(as.Date('2000-06-30', format='\%Y-\%m-\%d'), tz='etc/GMT+6')
p2 <- p2 + scale_x_datetime(date_breaks = "1 month",
limits=c(mintime, maxtime), date_labels = "\%b")
}
}
\author{
Kevin Shook
}
