% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNums.R, R/utils.R
\name{parseNums}
\alias{parseNums}
\title{Parses a string containing numbers}
\usage{
parseNums(numString)

parseNums(numString)
}
\arguments{
\item{numString}{Required. A character string containing numbers separated by any number of spaces.}
}
\value{
Returns a numeric vector.

Returns a numeric vector.
}
\description{
This is an internal \pkg{CRHMr} function and should normally not need to be called directly. It is used for reading data from messy files such as CRHM projects.

This is an internal \pkg{CRHMr} function and should normally not need to be called directly. It is used for reading data from messy files such as CRHM projects.
}
\examples{
parseNums(' 1  2 3   4     5 ')
parseNums(' 1  2 3   4     5 ')
}
\author{
Kevin Shook

Kevin Shook
}
\keyword{internal}
