% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qair2rh.R
\name{qair2rh}
\alias{qair2rh}
\title{qair2rh}
\usage{
qair2rh(qair, temp, press = 1013.25)
}
\arguments{
\item{qair}{specific humidity, dimensionless (e.g. kg/kg) ratio of water mass / total air mass}

\item{temp}{degrees C}

\item{press}{pressure in mb}
}
\value{
rh relative humidity, ratio of actual water mixing ratio to saturation mixing ratio
}
\description{
From Bolton 1980 The computation of Equivalent Potential Temperature
\url{http://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html}.
}
\details{
Convert specific humidity to relative humidity
}
\examples{
qair2rh(0.01, 10)
}
\author{
David LeBauer \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R}. Modified by Kevin Shook for air temps < 0 \eqn{^\circ}{ }C and to work for vectors or scalars.
}
