% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qair2rh.R
\name{qair2rh}
\alias{qair2rh}
\title{qair2rh}
\usage{
qair2rh(qair, temp, press = 101325, allowStupid = FALSE)
}
\arguments{
\item{qair}{Required. Specific humidity, dimensionless (e.g. kg/kg) ratio of water mass / total air mass.}

\item{temp}{Required. Air temp in degrees C.}

\item{press}{Optional. Air pressure in Pa. Default is 101325, the standard atmospheric air pressure.}

\item{allowStupid}{Optional. If \code{TRUE} then impossible values (smaller than zero or
greater than 1) will be allowed. The default value, \code{FALSE}, restricts the
returned value to be between 0 and 1.}
}
\value{
If successful, returns rh (relative humidity), the ratio of actual water
mixing ratio to saturation mixing ratio. Note that this value is a fraction (i.e.
between 0 and 1) rather than a percentage. If unsuccessful, returns the value \code{FALSE}.
}
\description{
From Bolton 1980 The computation of Equivalent Potential Temperature
\url{http://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html}.
}
\details{
Converts specific humidity to relative humidity
}
\examples{
qair2rh(0.0001, 10)
}
\references{
\url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R}.
}
\seealso{
\code{\link{qair2ea}}
}
\author{
David LeBauer. Modified by Kevin Shook for air temps < 0 \eqn{^\circ}{ }C and to work for vectors or scalars.
}
